mhdior {bayestestR}	R Documentation

Maximum HDI level inside/outside ROPE (MHDIOR)

Description

The MHDIOR (pronounced 'em-eich-dior') is an exploratory and non-validated index representing the maximum percentage of HDI that does not contain (or is entirely contained, in which case the value is prefixed with a negative sign), in the negligible values space defined by the ROPE. It differs from the ROPE percentage, i.e., from the proportion of a given CI in the ROPE, as it represents the maximum CI values needed to reach a ROPE proportion of 0% or 100%. Whether the index reflects the ROPE reaching 0% or 100% is indicated through the sign: a negative sign is added to indicate that the probability corresponds to the probability of a not significant effect (a percentage in ROPE of 100%). For instance, a MHDIOR of 97% means that there is a probability of .97 that a parameter (described by its posterior distribution) is outside the ROPE. In other words, the 97% HDI is the maximum HDI level for which the percentage in ROPE is 0%. On the contrary, a ROPE-based p of -97% indicates that there is a probability of .97 that the parameter is inside the ROPE (percentage in ROPE of 100%). A value close to 0% would indicate that the mode of the distribution falls perfectly at the edge of the ROPE, in which case the percentage of HDI needed to be on either side of the ROPE becomes infinitely small. Negative values do not refer to negative values per se, simply indicating that the value corresponds to non-significance rather than significance.

Usage

mhdior(x, ...)

## S3 method for class 'numeric'
mhdior(x, range = "default", precision = 0.1, ...)

## S3 method for class 'data.frame'
mhdior(x, range = "default", precision = 0.1, ...)

## S3 method for class 'emmGrid'
mhdior(x, range = "default", precision = 0.1, ...)

## S3 method for class 'BFBayesFactor'
mhdior(x, range = "default", precision = 0.1, ...)

## S3 method for class 'stanreg'
mhdior(
  x,
  range = "default",
  precision = 0.1,
  effects = c("fixed", "random", "all"),
  parameters = NULL,
  ...
)

## S3 method for class 'brmsfit'
mhdior(
  x,
  range = "default",
  precision = 0.1,
  effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL,
  ...
)

Arguments

x
Vector representing a posterior distribution. Can also be a stanreg or brmsfit model.

...
Currently not used.

range
ROPE's lower and higher bounds. Should be a vector of length two (e.g., c(-0.1, 0.1)) or "default". If "default", the range is set to c(-0.1, 0.1) if input is a vector, and based on rope_range() if a Bayesian model is provided.

precision
The precision by which to explore the ROPE space (in percentage). Lower values increase the precision of the returned p value but can be quite computationaly costly.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Examples

## Not run: 
library(bayestestR)

# precision = 1 is used to speed up examples...

mhdior(
  x = rnorm(1000, mean = 1, sd = 1),
  range = c(-0.1, 0.1),
  precision = 1
)

df <- data.frame(replicate(4, rnorm(100)))
mhdior(df, precision = 1)

if (require("rstanarm")) {
  model <- stan_glm(
    mpg ~ wt + gear, data = mtcars,
    chains = 2,
    iter = 200,
    refresh = 0
  )
  mhdior(model, precision = 1)
}

if (require("emmeans")) {
  mhdior(emtrends(model, ~1, "wt"))
}

if (require("brms")) {
  model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
  mhdior(model)
}

if (require("BayesFactor")) {
  bf <- ttestBF(x = rnorm(100, 1, 1))
  mhdior(bf)
}

## End(Not run)

[Package bayestestR version 0.7.0 Index]