Test for Practical Equivalence

Perform a Test for Practical Equivalence based on the "HDI+ROPE decision rule" (Kruschke 2018) to check whether parameter values should be accepted or rejected against an explicitely formulated "null hypothesis".

rope_test(posterior, bounds = c(-0.1, 0.1), CI = 90, verbose = TRUE)

Arguments

posterior
vector representing a posterior distribution.

bounds
ROPE's lower and higher bounds.

CI
the credible interval to use.

verbose
toggle off warnings.

Details

Using the ROPE and the HDI, Kruschke (2010, 2011, 2014, 2018) suggest using the percentage of the 95% (or 90%, considered more stable) HDI that falls within the ROPE as a decision rule. If the HDI is completely outside the ROPE, the "null hypothesis" for this parameter is "rejected". If the ROPE completely covers the HDI, i.e. all most credible values of a parameter are inside the region of practical equivalence, the null hypothesis is accepted. Else, it’s undecided whether to accept or reject the null hypothesis.

References

sjstats

Examples

library(bayestestR)

rope_test(posterior = rnorm(1000, 0, 0.01), bounds = c(-0.1, 0.1))

#> [1] "accepted"

rope_test(posterior = rnorm(1000, 0, 1), bounds = c(-0.1, 0.1))

#> [1] "undecided"

rope_test(posterior = rnorm(1000, 1, 0.01), bounds = c(-0.1, 0.1))

#> [1] "rejected"

rope_test(posterior = rnorm(1000, 1, 1), CI = c(50, 99))

#> $CI_50
#> [1] "rejected"
#> 
#> $CI_99
#> [1] "undecided"
#>