Inclusion Bayes Factors for effects across Bayesian models

Inclusion Bayes Factors for effects across Bayesian models

bayesfactor_inclusion(models, match_models = FALSE, prior_odds = NULL,
  ...)

Arguments

models
An object of class bayesfactor_models or BFBayesFactor.

match_models
If FALSE (default), Inclusion BFs are computed by comparing all models with an effect against all models without the effect. If TRUE, Inclusion BFs are computed by comparing all models with an effect against models without the effect AND without any higher-order interactions with the effect (additionally, interactions are compared only to models with the all main effects).

prior_odds
Optional vector of prior odds for the models. See priorOdds.

...
Arguments passed to or from other methods.

Value

a data frame containing the prior and posterior probabilities, and BF for each effect.

Details

Inclusion Bayes factors answer the question: Are the observed data more probable under models with a particular effect, than they are under models without that particular effect? In other words, on average - are models with effect \(X\) more likely to have produced the observed data than models without effect \(X\)?

See also the Bayes factors vignette.

Note

Random effects in the lme style will be displayed as interactions: i.e., (X|G) will become 1:G and X:G.

References

Hinne, M., Gronau, Q. F., van den Bergh, D., and Wagenmakers, E. (2019, March 25). A conceptual introduction to Bayesian Model Averaging. doi: 10.31234/osf.io/wgb64

Clyde, M. A., Ghosh, J., & Littman, M. L. (2011). Bayesian adaptive sampling for variable selection and model averaging. Journal of Computational and Graphical Statistics, 20(1), 80-101.

Mathot. S. (2017). Bayes like a Baws: Interpreting Bayesian Repeated Measures in JASP [Blog post]. Retrieved from https://www.cogsci.nl/blog/interpreting-bayesian-repeated-measures-in-jasp

Examples

library(bayestestR)

# Using bayesfactor_models:
# ------------------------------
mo0 <- lm(Sepal.Length ~ 1, data = iris)
mo1 <- lm(Sepal.Length ~ Species, data = iris)
mo2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
mo3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)

BFmodels <- bayesfactor_models(mo1, mo2, mo3, denominator = mo0)
bayesfactor_inclusion(BFmodels)

#>                      Pr(prior) Pr(posterior) Inclusion BF
#> Species                   0.75         1.000     2.02e+55
#> Petal.Length              0.50         1.000     3.58e+26
#> Species:Petal.Length      0.25         0.036         0.11
#> ---
#> Inclusion BFs compared among all models.

# BayesFactor
# -------------------------------
library(BayesFactor)

#> Loading required package: coda

#> Loading required package: Matrix

#> ************
#> Welcome to BayesFactor 0.9.12-4.2. If you have questions, please contact Richard Morey (richarddmorey@gmail.com).
#> 
#> Type BFManual() to open the manual.
#> ************

#> 
#> Attaching package: ‘BayesFactor’

#> The following object is masked from ‘package:testthat’:
#> 
#>     compare

BF <- generalTestBF(len ~ supp * dose, ToothGrowth, progress = FALSE)

bayesfactor_inclusion(BF)

#>           Pr(prior) Pr(posterior) Inclusion BF
#> supp            0.6          0.98        34.94
#> dose            0.6          1.00     5.73e+12
#> supp:dose       0.2          0.56         5.02
#> ---
#> Inclusion BFs compared among all models.

# compare only matched models:
bayesfactor_inclusion(BF, match_models = TRUE)

#>           Pr(prior) Pr(posterior) Inclusion BF
#> supp            0.4          0.42        22.67
#> dose            0.4          0.44     3.81e+12
#> supp:dose       0.2          0.56         1.31
#> ---
#> Inclusion BFs compared among matched models only.