Get inclusion BFs for effects across Bayesian models

Get inclusion BFs for effects across Bayesian models

bayesfactor_inclusion(models, match_models = FALSE, prior_odds = NULL)

Arguments

models
an object of class bayesfactor_models or BFBayesFactor.

match_models
If FALSE (default), Inclustion BFs are computed by comparing all models with an effect against all models without the effect. If TRUE, Inclusion BFs are computed by comparing all models with an effect against models without the effect AND without any higher-order interactions with the effect.

prior_odds
optional vector of prior odds for the models. See BayesFactor::priorOdds

Value

a data frame containing the prior and posterior probabilities, and BF (log) for each effect.

Note

Random effects in the lme style will be displayed as interactions: i.e., (X|G) will become 1:G and X:G.

References

Hinne, M., Gronau, Q. F., van den Bergh, D., & Wagenmakers, E. (2019, March 25). A conceptual introduction to Bayesian Model Averaging. https://doi.org/10.31234/osf.io/wgb64

Clyde, M. A., Ghosh, J., & Littman, M. L. (2011). Bayesian adaptive sampling for variable selection and model averaging. Journal of Computational and Graphical Statistics, 20(1), 80-101.

Mathot. S. (2017). Bayes like a Baws: Interpreting Bayesian Repeated Measures in JASP [Blog post]. Retrieved from https://www.cogsci.nl/blog/interpreting-bayesian-repeated-measures-in-jasp

Examples

library(bayestestR)

# Using with bayesfactor_models:
# ------------------------------
mo0 <- lm(Sepal.Length ~ 1, data = iris)
mo1 <- lm(Sepal.Length ~ Species, data = iris)
mo2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
mo3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)

BFmodels <- bayesfactor_models(mo1, mo2, mo3, denominator = mo0)
bayesfactor_inclusion(BFmodels)

#>                      Pr(prior) Pr(posterior) Inclusion.BF
#> Species                   0.75         1.000      2.0e+55
#> Petal.Length              0.50         1.000      3.6e+26
#> Species:Petal.Length      0.25         0.036      1.1e-01
#> ---
#> Inclusion BFs compared among all models.

# Using with BayesFactor objects:
# -------------------------------
library(BayesFactor)

#> Loading required package: coda

#> Loading required package: Matrix

#> ************
#> Welcome to BayesFactor 0.9.12-4.2. If you have questions, please contact Richard Morey (richarddmorey@gmail.com).
#> 
#> Type BFManual() to open the manual.
#> ************

#> 
#> Attaching package: 'BayesFactor'

#> The following object is masked from 'package:testthat':
#> 
#>     compare

BF <- generalTestBF(len ~ supp * dose, ToothGrowth, progress = FALSE)

bayesfactor_inclusion(BF)

#>           Pr(prior) Pr(posterior) Inclusion.BF
#> supp            0.6          0.98      3.5e+01
#> dose            0.6          1.00      5.7e+12
#> supp:dose       0.2          0.56      5.0e+00
#> ---
#> Inclusion BFs compared among all models.

# compare only matched models:
bayesfactor_inclusion(BF, match_models = TRUE)

#>           Pr(prior) Pr(posterior) Inclusion.BF
#> supp            0.4          0.42      2.3e+01
#> dose            0.4          0.44      3.8e+12
#> supp:dose       0.2          0.56      5.0e+00
#> ---
#> Inclusion BFs compared among matched models only.