Extract Bayes Factors from fitted models

This function computes Bayes factors from fitted models or extracts the Bayes factor from objects of class BFBayesFactor.

bayesfactor_models(..., denominator = 1)

Arguments

...
Fitted models (any models supported by insight), all fit on the same data, or a single BFBayesFactor object (see 'Details').

denominator
Either an integer indicating which of the models to use as the denominator, or a model to use as a denominator. Ignored for BFBayesFactor.

Value

A data frame containing the models' formulas (reconstructed fixed and random effects) and their BFs (log) of the supplied models, that prints nicely.

Details

For brmsfit or stanreg models, Bayes factors are computed using the bridgesampling package.

brmsfit models must have been fitted with save_all_pars = TRUE.

stanreg models must have been fitted with a defined diagnostic_file.

For BFBayesFactor, bayesfactor_models is mostly a wraparoud BayesFactor::extractBF.

For all other model types (supported by insight), BIC approximations are used to compute Bayes factors.

In order to correctly and precisely estimate Bayes Factors, a rule of thumb are the 4 P's: Proper Priors and Plentiful Posterior (i.e. probably at leat 40,000 samples instead of the default of 4,000).

References

Kass, R. E., & Raftery, A. E. (1995). Bayes Factors. Journal of the American Statistical Association, 90(430), 773-795.

Robert, C. P. (2016). The expected demise of the Bayes factor. Journal of Mathematical Psychology, 72, 33–37.

Wagenmakers, E. J. (2007). A practical solution to the pervasive problems of p values. Psychonomic bulletin & review, 14(5), 779-804.

Gronau, Q. F., Wagenmakers, E. J., Heck, D. W., & Matzke, D. (2019). A simple method for comparing complex models: Bayesian model comparison for hierarchical multinomial processing tree models using Warp-III bridge sampling. Psychometrika, 84(1), 261-284.

See also

Examples

# With lm objects:
lm1 <- lm(Sepal.Length ~ 1, data = iris)
lm2 <- lm(Sepal.Length ~ Species, data = iris)
lm3 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
lm4 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
bayesfactor_models(lm1, lm2, lm3, lm4, denominator = 1)

#> Bayes factor analysis
#> --------------                                     
#> [2] Species                   1.7e+29
#> [3] Species + Petal.Length    5.8e+55
#> [4] Species * Petal.Length    2.2e+54
#> 
#> Against denominator:
#> 	 [1] (Intercept only)   
#> ---
#> Bayes factor type:  BIC approximation 

bayesfactor_models(lm2, lm3, lm4, denominator = lm1) # same result

#> Bayes factor analysis
#> --------------                                     
#> [1] Species                   1.7e+29
#> [2] Species + Petal.Length    5.8e+55
#> [3] Species * Petal.Length    2.2e+54
#> 
#> Against denominator:
#> 	 [4] (Intercept only)   
#> ---
#> Bayes factor type:  BIC approximation 

bayesfactor_models(lm1, lm2, lm3, lm4, denominator = lm1) # same result

#> Bayes factor analysis
#> --------------                                     
#> [2] Species                   1.7e+29
#> [3] Species + Petal.Length    5.8e+55
#> [4] Species * Petal.Length    2.2e+54
#> 
#> Against denominator:
#> 	 [1] (Intercept only)   
#> ---
#> Bayes factor type:  BIC approximation 

# With lmerMod objects:
library(lme4)
lmer1 <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
lmer2 <- lmer(Sepal.Length ~ Petal.Length + (Petal.Length | Species), data = iris)

#> boundary (singular) fit: see ?isSingular

lmer3 <- lmer(
  Sepal.Length ~ Petal.Length + (Petal.Length | Species) + (1 | Petal.Width),
  data = iris
)

#> boundary (singular) fit: see ?isSingular

bayesfactor_models(lmer1, lmer2, lmer3, denominator = 1)

#> Bayes factor analysis
#> --------------                                                                         
#> [2] Petal.Length + (Petal.Length | Species)                        0.0579
#> [3] Petal.Length + (Petal.Length | Species) + (1 | Petal.Width)    0.0047
#> 
#> Against denominator:
#> 	 [1] Petal.Length + (1 | Species)   
#> ---
#> Bayes factor type:  BIC approximation 

bayesfactor_models(lmer1, lmer2, lmer3, denominator = lmer1)

#> Bayes factor analysis
#> --------------                                                                         
#> [2] Petal.Length + (Petal.Length | Species)                        0.0579
#> [3] Petal.Length + (Petal.Length | Species) + (1 | Petal.Width)    0.0047
#> 
#> Against denominator:
#> 	 [1] Petal.Length + (1 | Species)   
#> ---
#> Bayes factor type:  BIC approximation 
# NOT RUN {
# With BFBayesFactor objects:
library(BayesFactor)
data(puzzles)
BF <- anovaBF(RT ~ shape * color + ID,
  data = puzzles,
  whichRandom = "ID", progress = FALSE
)
BF
bayesfactor_models(BF) # basically the same

# with brmfit objects:
# (note the save_all_pars MUST be set to TRUE in order to work)
library(brms)
brm1 <- brm(Sepal.Length ~ 1, data = iris, save_all_pars = TRUE)
brm2 <- brm(Sepal.Length ~ Species, data = iris, save_all_pars = TRUE)
brm3 <- brm(
  Sepal.Length ~ Species + Petal.Length,
  data = iris,
  save_all_pars = TRUE
)

bayesfactor_models(brm1, brm2, brm3, denominator = 1)
# }