Savage-Dickey density ratio Bayes Factor (BF)

This method computes Bayes factors against the null (either a point or an interval), bases on prior and posterior samples of a single parameter. This Bayes factor indicates the degree by which the mass of the posterior distribution has shifted further away from or closer to the null value(s) (relative to the prior distribution), thus indicating if the null value has become less or more likely given the observed data.

When the null is an interval, the Bayes factor is computed by comparing the prior and posterior odds of the parameter falling within or outside the null; When the null is a point, a Savage-Dickey density ratio is computed, which is also an approximation of a Bayes factor comparing the marginal likelihoods of the model against a model in which the tested parameter has been restricted to the point null.

For info on specifying correct priors for factors with more than 2 levels, see the Bayes factors vignette.

For more info, see the Bayes factors vignette.

bayesfactor_parameters(posterior, prior = NULL,
  direction = "two-sided", null = 0, verbose = TRUE, ...)

bayesfactor_savagedickey(posterior, prior = NULL,
  direction = "two-sided", null = 0, verbose = TRUE,
  hypothesis = NULL, ...)

# S3 method for numeric
bayesfactor_parameters(posterior, prior = NULL,
  direction = "two-sided", null = 0, verbose = TRUE, ...)

# S3 method for stanreg
bayesfactor_parameters(posterior, prior = NULL,
  direction = "two-sided", null = 0, verbose = TRUE,
  effects = c("fixed", "random", "all"), component = c("conditional",
  "zi", "zero_inflated", "all"), ...)

# S3 method for brmsfit
bayesfactor_parameters(posterior, prior = NULL,
  direction = "two-sided", null = 0, verbose = TRUE,
  effects = c("fixed", "random", "all"), component = c("conditional",
  "zi", "zero_inflated", "all"), ...)

# S3 method for emmGrid
bayesfactor_parameters(posterior, prior = NULL,
  direction = "two-sided", null = 0, verbose = TRUE, ...)

# S3 method for data.frame
bayesfactor_parameters(posterior, prior = NULL,
  direction = "two-sided", null = 0, verbose = TRUE, ...)

Arguments

posterior
A numerical vector, stanreg / brmsfit object, emmGrid or a data frame - representing a posterior distribution(s) from (see Details).

prior
An object representing a prior distribution (see Details).

direction
Test type (see details). One of 0, "two-sided" (default, two tailed), -1, "left" (left tailed) or 1, "right" (right tailed).

null
Value of the null, either a scaler (for point-null) or a a range (for a interval-null).

verbose
Toggle off warnings.

...
Currently not used.

hypothesis
Deprecated in favour of null.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Value

A data frame containing the Bayes factor representing evidence against the null.

Details

This method is used to compute Bayes factors based on prior and posterior distributions. When posterior is a model (stanreg, brmsfit), posterior and prior samples are extracted for each parameter, and Savage-Dickey Bayes factors are computed for each parameter.

NOTE: For brmsfit models, the model must have been fitted with custom (non-default) priors. See example below.

Setting the correct prior
It is important to provide the correct prior for meaningful results.

When posterior is a numerical vector, prior should also be a numerical vector.

When posterior is a data.frame, prior should also be a data.frame, with matching column order.

When posterior is a stanreg or brmsfit model:

prior can be set to NULL, in which case prior samples are drawn internally.

prior can also be a model equvilant to posterior but with samples from the priors only.

When posterior is an emmGrid object:

prior should be the stanreg or brmsfit model used to create the emmGrid objects.

prior can also be an emmGrid object equvilant to posterior but created with a model of priors samples only.

One-sided Tests (setting an order restriction)
One sided tests (controlled by direction) are conducted by restricting the prior and posterior of the non-null values (the "alternative") to one side of the null only (Morey & Wagenmakers, 2013). For example, if we have a prior hypothesis that the parameter should be positive, the alternative will be restricted to the region to the right of the null (point or interval).
Interpreting Bayes Factors
A Bayes factor greater than 1 can be interpereted as evidence against the null, at which one convention is that a Bayes factor greater than 3 can be considered as "substantial" evidence against the null (and vice versa, a Bayes factor smaller than 1/3 indicates substantial evidence in favor of the null-model) (Wetzels et al. 2011).
References

Wagenmakers, E. J., Lodewyckx, T., Kuriyal, H., and Grasman, R. (2010). Bayesian hypothesis testing for psychologists: A tutorial on the Savage-Dickey method. Cognitive psychology, 60(3), 158-189.

Wetzels, R., Matzke, D., Lee, M. D., Rouder, J. N., Iverson, G. J., and Wagenmakers, E.-J. (2011). Statistical Evidence in Experimental Psychology: An Empirical Comparison Using 855 t Tests. Perspectives on Psychological Science, 6(3), 291–298. doi: 10.1177/1745691611406923

Heck, D. W. (2019). A caveat on the Savage–Dickey density ratio: The case of computing Bayes factors for regression parameters. British Journal of Mathematical and Statistical Psychology, 72(2), 316-333.

Morey, R. D., & Wagenmakers, E. J. (2014). Simple relation between Bayesian order-restricted and point-null hypothesis tests. Statistics & Probability Letters, 92, 121-124.

Morey, R. D., & Rouder, J. N. (2011). Bayes factor approaches for testing interval null hypotheses. Psychological methods, 16(4), 406.

Examples

library(bayestestR)

prior <- distribution_normal(1000, mean = 0, sd = 1)
posterior <- distribution_normal(1000, mean = .5, sd = .3)

bayesfactor_parameters(posterior, prior)

#> # Bayes Factor (Savage-Dickey density ratio)
#> 
#>  Bayes Factor
#>          1.23
#> 
#> * Evidence Against The Null: [0]
#> 
# NOT RUN {
# rstanarm models
# ---------------
library(rstanarm)
contrasts(sleep$group) <- contr.bayes # see vingette
stan_model <- stan_lmer(extra ~ group + (1 | ID), data = sleep)
bayesfactor_parameters(stan_model)
bayesfactor_parameters(stan_model, null = rope_range(stan_model))

# emmGrid objects
# ---------------
library(emmeans)
group_diff <- pairs(emmeans(stan_model, ~group))
bayesfactor_parameters(group_diff, prior = stan_model)

# brms models
# -----------
library(brms)
contrasts(sleep$group) <- contr.bayes # see vingette
my_custom_priors <-
  set_prior("student_t(3, 0, 1)", class = "b") +
  set_prior("student_t(3, 0, 1)", class = "sd", group = "ID")

brms_model <- brm(extra ~ group + (1 | ID),
  data = sleep,
  prior = my_custom_priors
)
bayesfactor_parameters(brms_model)
# }