bayestestR-package {bayestestR}	R Documentation

bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework

Description

Existing R packages allow users to easily fit a large variety of models and extract and visualize the posterior draws. However, most of these packages only return a limited set of indices (e.g., point-estimates and CIs). bayestestR provides a comprehensive and consistent set of functions to analyze and describe posterior distributions generated by a variety of models objects, including popular modeling packages such as rstanarm, brms or BayesFactor.

References:

Makowski et al. (2019) doi: 10.21105/joss.01541

Makowski et al. (2019) doi: 10.3389/fpsyg.2019.02767

Details

bayestestR

Author(s)

Maintainer: Dominique Makowski dom.makowski@gmail.com (ORCID) (@Dom_Makowski)

Authors:

Daniel Lüdecke d.luedecke@uke.de (ORCID) (@strengejacke)

Mattan S. Ben-Shachar matanshm@post.bgu.ac.il (ORCID) (@mattansb)

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID) (@patilindrajeets)

Michael D. Wilson michael.d.wilson@curtin.edu.au (ORCID)

Brenton M. Wiernik brenton@wiernik.org (ORCID) (@bmwiernik)

Other contributors:

Paul-Christian Bürkner paul.buerkner@gmail.com [reviewer]

Tristan Mahr tristan.mahr@wisc.edu (ORCID) [reviewer]

Henrik Singmann singmann@gmail.com (ORCID) [contributor]

Quentin F. Gronau (ORCID) [contributor]

Sam Crawley sam@crawley.nz (ORCID) [contributor]

See Also

Useful links:

https://easystats.github.io/bayestestR/

Report bugs at https://github.com/easystats/bayestestR/issues

[Package bayestestR version 0.13.2 Index]