Confidence/Credible/Compatibility Interval (CI)

Compute Confidence/Credible/Compatibility Intervals (CI) or Support Intervals (SI) for Bayesian and frequentist models. The Documentation is accessible for:

Usage

ci(x, ...)

# S3 method for numeric
ci(x, ci = 0.95, method = "ETI", verbose = TRUE, BF = 1, ...)

# S3 method for data.frame
ci(x, ci = 0.95, method = "ETI", verbose = TRUE, BF = 1, ...)

# S3 method for sim.merMod
ci(
  x,
  ci = 0.95,
  method = "ETI",
  effects = c("fixed", "random", "all"),
  parameters = NULL,
  verbose = TRUE,
  ...
)

# S3 method for sim
ci(x, ci = 0.95, method = "ETI", parameters = NULL, verbose = TRUE, ...)

# S3 method for stanreg
ci(
  x,
  ci = 0.95,
  method = "ETI",
  effects = c("fixed", "random", "all"),
  component = c("location", "all", "conditional", "smooth_terms", "sigma",
    "distributional", "auxiliary"),
  parameters = NULL,
  verbose = TRUE,
  BF = 1,
  ...
)

# S3 method for brmsfit
ci(
  x,
  ci = 0.95,
  method = "ETI",
  effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL,
  verbose = TRUE,
  BF = 1,
  ...
)

# S3 method for BFBayesFactor
ci(x, ci = 0.95, method = "ETI", verbose = TRUE, BF = 1, ...)

# S3 method for MCMCglmm
ci(x, ci = 0.95, method = "ETI", verbose = TRUE, ...)

Arguments

x

A stanreg or brmsfit model, or a vector representing a posterior distribution.

...

Currently not used.

ci

Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .95 (95%).

method

Can be 'ETI' (default), 'HDI', 'BCI', 'SPI' or 'SI'.

verbose

Toggle off warnings.

BF

The amount of support required to be included in the support interval.

effects

Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters

Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component

Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Value

A data frame with following columns:

Parameter The model parameter(s), if x is a model-object. If x is a vector, this column is missing.

CI The probability of the credible interval.

CI_low, CI_high The lower and upper credible interval limits for the parameters.

Details

Bayesian models

Frequentist models

Note

When it comes to interpretation, we recommend thinking of the CI in terms of an "uncertainty" or "compatibility" interval, the latter being defined as “Given any value in the interval and the background assumptions, the data should not seem very surprising” (Gelman & Greenland 2019).

There is also a plot()-method implemented in the see-package.

References

Gelman A, Greenland S. Are confidence intervals better termed "uncertainty intervals"? BMJ 2019;l5381. 10.1136/bmj.l5381

See also

Examples

library(bayestestR)

posterior <- rnorm(1000)
ci(posterior, method = "ETI")
#> 95% ETI: [-2.06, 2.04]
ci(posterior, method = "HDI")
#> 95% HDI: [-1.90, 2.14]

df <- data.frame(replicate(4, rnorm(100)))
ci(df, method = "ETI", ci = c(0.80, 0.89, 0.95))
#> Equal-Tailed Interval
#> 
#> Parameter |       80% ETI |       89% ETI |       95% ETI
#> ---------------------------------------------------------
#> X1        | [-1.37, 1.29] | [-1.62, 1.55] | [-2.15, 1.84]
#> X2        | [-1.03, 1.18] | [-1.34, 1.40] | [-1.61, 1.94]
#> X3        | [-1.41, 1.40] | [-1.69, 1.59] | [-1.80, 1.91]
#> X4        | [-1.23, 1.21] | [-1.64, 1.49] | [-2.43, 1.85]
ci(df, method = "HDI", ci = c(0.80, 0.89, 0.95))
#> Highest Density Interval
#> 
#> Parameter |       80% HDI |       89% HDI |       95% HDI
#> ---------------------------------------------------------
#> X1        | [-1.04, 1.56] | [-1.48, 1.64] | [-1.79, 2.32]
#> X2        | [-1.13, 1.04] | [-1.28, 1.53] | [-1.64, 1.96]
#> X3        | [-1.87, 0.70] | [-1.74, 1.51] | [-1.87, 1.66]
#> X4        | [-1.04, 1.39] | [-1.53, 1.65] | [-2.11, 1.97]

model <- stan_glm(mpg ~ wt, data = mtcars, chains = 2, iter = 200, refresh = 0)
#> Warning: The largest R-hat is 1.06, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
ci(model, method = "ETI", ci = c(0.80, 0.89))
#> Equal-Tailed Interval
#> 
#> Parameter   |        80% ETI |        89% ETI | Effects |   Component
#> ---------------------------------------------------------------------
#> (Intercept) | [34.13, 39.79] | [33.43, 40.64] |   fixed | conditional
#> wt          | [-6.12, -4.46] | [-6.36, -4.23] |   fixed | conditional
ci(model, method = "HDI", ci = c(0.80, 0.89))
#> Highest Density Interval 
#> 
#> Parameter   |        80% HDI |        89% HDI
#> ---------------------------------------------
#> (Intercept) | [34.72, 40.05] | [33.58, 40.69]
#> wt          | [-6.11, -4.44] | [-6.37, -4.23]
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
#> Compiling Stan program...
#> Start sampling
#> 
#> SAMPLING FOR MODEL '667cda6873b3c0bde53ba0b20ec47308' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.13 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.025958 seconds (Warm-up)
#> Chain 1:                0.025139 seconds (Sampling)
#> Chain 1:                0.051097 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL '667cda6873b3c0bde53ba0b20ec47308' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 1e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.1 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.026103 seconds (Warm-up)
#> Chain 2:                0.022128 seconds (Sampling)
#> Chain 2:                0.048231 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL '667cda6873b3c0bde53ba0b20ec47308' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 8e-06 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.08 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.025469 seconds (Warm-up)
#> Chain 3:                0.023692 seconds (Sampling)
#> Chain 3:                0.049161 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL '667cda6873b3c0bde53ba0b20ec47308' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 7e-06 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.07 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.025833 seconds (Warm-up)
#> Chain 4:                0.027995 seconds (Sampling)
#> Chain 4:                0.053828 seconds (Total)
#> Chain 4: 
ci(model, method = "ETI")
#> Equal-Tailed Interval
#> 
#> Parameter   |        95% ETI | Effects |   Component
#> ----------------------------------------------------
#> b_Intercept | [36.19, 43.31] |   fixed | conditional
#> b_wt        | [-4.80, -1.62] |   fixed | conditional
#> b_cyl       | [-2.37, -0.64] |   fixed | conditional
ci(model, method = "HDI")
#> Highest Density Interval 
#> 
#> Parameter   |        95% HDI
#> ----------------------------
#> (Intercept) | [36.00, 43.08]
#> wt          | [-4.78, -1.60]
#> cyl         | [-2.37, -0.64]
bf <- ttestBF(x = rnorm(100, 1, 1))
ci(bf, method = "ETI")
#> Equal-Tailed Interval
#> 
#> Parameter  |      95% ETI
#> -------------------------
#> Difference | [0.80, 1.18]
ci(bf, method = "HDI")
#> Highest Density Interval
#> 
#> Parameter  |      95% HDI
#> -------------------------
#> Difference | [0.79, 1.18]
model <- emtrends(model, ~1, "wt")
ci(model, method = "ETI")
#> Equal-Tailed Interval
#> 
#> Parameter |        95% ETI
#> --------------------------
#> overall   | [-4.80, -1.62]
ci(model, method = "HDI")
#> Highest Density Interval
#> 
#> Parameter |        95% HDI
#> --------------------------
#> overall   | [-4.78, -1.60]