Describe Posterior Distributions

Compute indices relevant to describe and characterize the posterior distributions.

describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  ...
)

# S3 method for numeric
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  bf_prior = NULL,
  BF = 1,
  ...
)

# S3 method for stanreg
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"),
  priors = FALSE,
  effects = c("fixed", "random", "all"),
  parameters = NULL,
  BF = 1,
  ...
)

# S3 method for stanmvreg
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = "p_direction",
  rope_range = "default",
  rope_ci = 0.89,
  bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"),
  priors = FALSE,
  effects = c("fixed", "random", "all"),
  parameters = NULL,
  ...
)

# S3 method for MCMCglmm
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  diagnostic = "ESS",
  parameters = NULL,
  ...
)

# S3 method for brmsfit
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"),
  effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL,
  BF = 1,
  ...
)

# S3 method for BFBayesFactor
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope", "bf"),
  rope_range = "default",
  rope_ci = 0.89,
  priors = TRUE,
  ...
)

Arguments

posteriors
A vector, data frame or model of posterior draws.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi), "ETI" (see eti) or "SI" (see si).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

...
Additional arguments to be passed to or from methods.

bf_prior
Distribution representing a prior for the computation of Bayes factors / SI. Used if the input is a posterior, otherwise (in the case of models) ignored.

BF
The amount of support required to be included in the support interval.

diagnostic
Diagnostic metrics to compute. Character (vector) or list with one or more of these options: "ESS", "Rhat", "MCSE" or "all".

priors
Add the prior used for each parameter.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Details

One or more components of point estimates (like posterior mean or median), intervals and tests can be omitted from the summary output by setting the related argument to NULL. For example, test = NULL and centrality = NULL would only return the HDI (or CI).

References

Comparison of Point-Estimates

Region of Practical Equivalence (ROPE)

Bayes factors

Examples

library(bayestestR)

x <- rnorm(1000)
describe_posterior(x)

#>   Parameter     Median CI    CI_low  CI_high    pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior 0.03343118 89 -1.541518 1.552591 0.518      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.08641975

describe_posterior(x, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify a prior (in the form 'prior = distribution_normal(1000, 0, 1)') to get meaningful results.

#>   Parameter     Median      MAD       Mean        SD          MAP CI    CI_low
#> 1 Posterior 0.03343118 1.029941 0.04029485 0.9846458 -0.004019596 89 -1.541518
#>    CI_high     p_map    pd p_ROPE    ps ROPE_CI ROPE_low ROPE_high
#> 1 1.552591 0.9999988 0.518  0.077 0.472      89     -0.1       0.1
#>   ROPE_Percentage ROPE_Equivalence BF
#> 1      0.08641975        Undecided  1

describe_posterior(x, ci = c(0.80, 0.90))

#>   Parameter     Median CI    CI_low  CI_high    pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior 0.03343118 80 -1.284912 1.204892 0.518      89     -0.1       0.1
#> 2 Posterior 0.03343118 90 -1.624386 1.559000 0.518      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.08641975
#> 2      0.08641975

df <- data.frame(replicate(4, rnorm(100)))
describe_posterior(df)

#>   Parameter      Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1        X1  0.06569003 89 -1.510397 1.272253 0.51      89     -0.1       0.1
#> 2        X2  0.12795196 89 -1.591547 1.320824 0.51      89     -0.1       0.1
#> 3        X3 -0.07166458 89 -2.005170 1.200516 0.54      89     -0.1       0.1
#> 4        X4  0.11840365 89 -1.240517 1.491279 0.57      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.04444444
#> 2      0.04444444
#> 3      0.10000000
#> 4      0.07777778

describe_posterior(df, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify priors (with column order matching 'posterior') to get meaningful results.

#>   Parameter      Median       MAD        Mean        SD         MAP CI
#> 1        X1  0.06569003 0.9529887 -0.02998941 0.9198598  0.24226295 89
#> 2        X2  0.12795196 1.0941930  0.10101597 1.0241139 -0.06245219 89
#> 3        X3 -0.07166458 0.9374214 -0.16503654 0.9878320  0.10804210 89
#> 4        X4  0.11840365 1.0070136  0.13744968 0.9142465  0.57653501 89
#>      CI_low  CI_high     p_MAP   pd p_ROPE   ps ROPE_CI ROPE_low ROPE_high
#> 1 -1.510397 1.272253 0.9624055 0.51   0.04 0.50      89     -0.1       0.1
#> 2 -1.591547 1.320824 0.9986274 0.51   0.04 0.51      89     -0.1       0.1
#> 3 -2.005170 1.200516 0.9906335 0.54   0.09 0.49      89     -0.1       0.1
#> 4 -1.240517 1.491279 0.9901760 0.57   0.07 0.52      89     -0.1       0.1
#>   ROPE_Percentage ROPE_Equivalence BF
#> 1      0.04444444        Undecided  1
#> 2      0.04444444        Undecided  1
#> 3      0.10000000        Undecided  1
#> 4      0.07777778        Undecided  1

describe_posterior(df, ci = c(0.80, 0.90))

#>   Parameter      Median CI     CI_low   CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1        X1  0.06569003 80 -1.3114222 0.8763943 0.51      89     -0.1       0.1
#> 2        X1  0.06569003 90 -1.5103968 1.3851220 0.51      89     -0.1       0.1
#> 3        X2  0.12795196 80 -0.9219137 1.3208244 0.51      89     -0.1       0.1
#> 4        X2  0.12795196 90 -1.6649464 1.3208244 0.51      89     -0.1       0.1
#> 5        X3 -0.07166458 80 -1.4715541 0.9924637 0.54      89     -0.1       0.1
#> 6        X3 -0.07166458 90 -2.1587528 1.0887519 0.54      89     -0.1       0.1
#> 7        X4  0.11840365 80 -1.2040960 1.0847977 0.57      89     -0.1       0.1
#> 8        X4  0.11840365 90 -1.4280646 1.4726549 0.57      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.04444444
#> 2      0.04444444
#> 3      0.04444444
#> 4      0.04444444
#> 5      0.10000000
#> 6      0.10000000
#> 7      0.07777778
#> 8      0.07777778

if (FALSE) {
# rstanarm models
# -----------------------------------------------
if (require("rstanarm") && require("emmeans")) {
  model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200, refresh = 0)
  describe_posterior(model)
  describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")
  describe_posterior(model, ci = c(0.80, 0.90))

  # emmeans estimates
  # -----------------------------------------------
  describe_posterior(emtrends(model, ~1, "wt"))
}

# brms models
# -----------------------------------------------
if (require("brms")) {
  model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
  describe_posterior(model)
  describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")
  describe_posterior(model, ci = c(0.80, 0.90))
}

# BayesFactor objects
# -----------------------------------------------
if (require("BayesFactor")) {
  bf <- ttestBF(x = rnorm(100, 1, 1))
  describe_posterior(bf)
  describe_posterior(bf, centrality = "all", dispersion = TRUE, test = "all")
  describe_posterior(bf, ci = c(0.80, 0.90))
}
}