Describe Posterior Distributions

Compute indices relevant to describe and characterise the posterior distributions.

describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, ...)

# S3 method for numeric
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, ...)

# S3 method for stanreg
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, diagnostic = c("ESS", "Rhat"),
  priors = TRUE, effects = c("fixed", "random", "all"),
  parameters = NULL, ...)

# S3 method for brmsfit
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, diagnostic = c("ESS", "Rhat"),
  effects = c("fixed", "random", "all"), component = c("conditional",
  "zi", "zero_inflated", "all"), parameters = NULL, ...)

# S3 method for BFBayesFactor
describe_posterior(posteriors,
  centrality = "median", dispersion = FALSE, ci = 0.89,
  ci_method = "hdi", test = c("p_direction", "rope", "bf"),
  rope_range = "default", rope_ci = 0.89, priors = TRUE, ...)

Arguments

posteriors
A vector, dataframe or model of posterior draws.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi) or "ETI" (see eti).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests.#' For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

...
Additional arguments to be passed to or from methods.

bf_prior
Distribution representing a prior for the computation of Bayes factors. Used if the input is a posterior, otherwise (in the case of models) ignored.

diagnostic
Diagnostic metrics to compute. Character (vector) or list with one or more of these options: "ESS", "Rhat", "MCSE" or "all".

priors
Add the prior used for each parameter.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Details

One or more components of point estimates (like posterior mean or median), intervals and tests can be ommitted from the summary output by setting the related argument to NULL. For example, test = NULL and centrality = NULL would only return the HDI (or CI).

References

Comparison of Point-Estimates

Region of Practical Equivalence (ROPE)

Bayes factors

Examples

library(bayestestR)

x <- rnorm(1000)
describe_posterior(x)

#>   Parameter    Median CI    CI_low  CI_high    pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior 0.0336808 89 -1.689547 1.516557 0.515      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.09090909

describe_posterior(x, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify a prior (in the form 'prior = distribution_normal(1000, 0, 1)') to get meaningful results.

#>   Parameter    Median      MAD       Mean       SD        MAP CI    CI_low
#> 1 Posterior 0.0336808 1.013192 0.03673998 1.004366 0.02488973 89 -1.689547
#>    CI_high     p_map    pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage
#> 1 1.516557 0.9995711 0.515      89     -0.1       0.1      0.09090909
#>   ROPE_Equivalence BF
#> 1        Undecided  1

describe_posterior(x, ci = c(0.80, 0.90))

#>   Parameter Median CI CI_low CI_high pd ROPE_CI ROPE_low ROPE_high
#> 1      <NA>     NA NA     NA      NA NA      NA       NA        NA
#> 2      <NA>     NA NA     NA      NA NA      NA       NA        NA
#>   ROPE_Percentage
#> 1              NA
#> 2              NA

df <- data.frame(replicate(4, rnorm(100)))
describe_posterior(df)

#>   Parameter      Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1        X1  0.03683574 89 -1.792781 1.311301 0.52      89     -0.1       0.1
#> 2        X2  0.02898097 89 -1.485226 1.113986 0.52      89     -0.1       0.1
#> 3        X3 -0.07666456 89 -1.587907 1.661350 0.51      89     -0.1       0.1
#> 4        X4 -0.17897730 89 -1.325198 1.515028 0.63      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.10000000
#> 2      0.14444444
#> 3      0.04444444
#> 4      0.07777778

describe_posterior(df, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify priors (with column order matching 'posterior') to get meaningful results.

#>   Parameter      Median       MAD         Mean        SD         MAP CI
#> 1        X1  0.03683574 0.9125343 -0.009615073 0.9683536  0.27275479 89
#> 2        X2  0.02898097 0.8114604  0.083259751 0.9016977  0.11428834 89
#> 3        X3 -0.07666456 1.0145283 -0.081999382 1.0115371  0.08490191 89
#> 4        X4 -0.17897730 0.8015870 -0.203107107 0.9222455 -0.26763522 89
#>      CI_low  CI_high     p_MAP   pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage
#> 1 -1.792781 1.311301 0.9655611 0.52      89     -0.1       0.1      0.10000000
#> 2 -1.485226 1.113986 0.9879746 0.52      89     -0.1       0.1      0.14444444
#> 3 -1.587907 1.661350 0.9982497 0.51      89     -0.1       0.1      0.04444444
#> 4 -1.325198 1.515028 0.9180627 0.63      89     -0.1       0.1      0.07777778
#>   ROPE_Equivalence BF
#> 1        Undecided  1
#> 2        Undecided  1
#> 3        Undecided  1
#> 4        Undecided  1

describe_posterior(df, ci = c(0.80, 0.90))

#>   Parameter      Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1        X1  0.03683574 80 -1.339659 0.977814 0.52      89     -0.1       0.1
#> 2        X1  0.03683574 90 -1.792781 1.372876 0.52      89     -0.1       0.1
#> 3        X2  0.02898097 80 -1.151562 0.910437 0.52      89     -0.1       0.1
#> 4        X2  0.02898097 90 -1.591667 1.113986 0.52      89     -0.1       0.1
#> 5        X3 -0.07666456 80 -1.224793 1.275265 0.51      89     -0.1       0.1
#> 6        X3 -0.07666456 90 -1.587907 1.721427 0.51      89     -0.1       0.1
#> 7        X4 -0.17897730 80 -1.220785 1.014600 0.63      89     -0.1       0.1
#> 8        X4 -0.17897730 90 -1.445853 1.515028 0.63      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.10000000
#> 2      0.10000000
#> 3      0.14444444
#> 4      0.14444444
#> 5      0.04444444
#> 6      0.04444444
#> 7      0.07777778
#> 8      0.07777778

# rstanarm models
# -----------------------------------------------
library(rstanarm)
model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200, refresh = 0)
describe_posterior(model)

#>     Parameter     Median CI    CI_low   CI_high    pd ROPE_CI   ROPE_low
#> 1 (Intercept) 39.4499167 89 31.565557 47.566166 1.000      89 -0.6026948
#> 2          wt -5.4817002 89 -6.718247 -4.399036 1.000      89 -0.6026948
#> 3        gear -0.3422758 89 -1.695148  1.100377 0.635      89 -0.6026948
#>   ROPE_high ROPE_Percentage ESS     Rhat Prior_Distribution Prior_Location
#> 1 0.6026948       0.0000000 212 1.000945             normal              0
#> 2 0.6026948       0.0000000 223 1.000373             normal              0
#> 3 0.6026948       0.4748603 233 1.001920             normal              0
#>   Prior_Scale
#> 1    60.26948
#> 2    15.39911
#> 3    20.42191

describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")

#> Computation of Bayes factors: sampling priors, please wait...

#>     Parameter     Median       MAD       Mean        SD        MAP CI    CI_low
#> 1 (Intercept) 39.4499167 5.3638837 39.0397579 5.1245289 40.2974904 89 31.565557
#> 2          wt -5.4817002 0.7376005 -5.5224993 0.7636675 -5.3897288 89 -6.718247
#> 3        gear -0.3422758 0.9680601 -0.3180788 0.9213700 -0.5520105 89 -1.695148
#>     CI_high     p_MAP    pd ROPE_CI   ROPE_low ROPE_high ROPE_Percentage
#> 1 47.566166 0.0000000 1.000      89 -0.6026948 0.6026948       0.0000000
#> 2 -4.399036 0.0000000 1.000      89 -0.6026948 0.6026948       0.0000000
#> 3  1.100377 0.8877973 0.635      89 -0.6026948 0.6026948       0.4748603
#>   ROPE_Equivalence           BF ESS     Rhat Prior_Distribution Prior_Location
#> 1         Rejected 2.959977e+04 212 1.000945             normal              0
#> 2         Rejected 2.655481e+04 223 1.000373             normal              0
#> 3        Undecided 5.536644e-02 233 1.001920             normal              0
#>   Prior_Scale
#> 1    60.26948
#> 2    15.39911
#> 3    20.42191

describe_posterior(model, ci = c(0.80, 0.90))

#>     Parameter     Median CI    CI_low    CI_high    pd ROPE_CI   ROPE_low
#> 1 (Intercept) 39.4499167 80 32.029803 44.7142675 1.000      89 -0.6026948
#> 2 (Intercept) 39.4499167 90 29.667488 46.3935618 1.000      89 -0.6026948
#> 3          wt -5.4817002 80 -6.384793 -4.4963720 1.000      89 -0.6026948
#> 4          wt -5.4817002 90 -6.689700 -4.2429754 1.000      89 -0.6026948
#> 5        gear -0.3422758 80 -1.256913  0.9985041 0.635      89 -0.6026948
#> 6        gear -0.3422758 90 -1.695148  1.2540259 0.635      89 -0.6026948
#>   ROPE_high ROPE_Percentage ESS     Rhat Prior_Distribution Prior_Location
#> 1 0.6026948       0.0000000 212 1.000945             normal              0
#> 2 0.6026948       0.0000000 212 1.000945             normal              0
#> 3 0.6026948       0.0000000 223 1.000373             normal              0
#> 4 0.6026948       0.0000000 223 1.000373             normal              0
#> 5 0.6026948       0.4748603 233 1.001920             normal              0
#> 6 0.6026948       0.4748603 233 1.001920             normal              0
#>   Prior_Scale
#> 1    60.26948
#> 2    60.26948
#> 3    15.39911
#> 4    15.39911
#> 5    20.42191
#> 6    20.42191

# emmeans estimates
# -----------------------------------------------
library(emmeans)
describe_posterior(emtrends(model, ~1, "wt"))

#>   Parameter  Median CI    CI_low   CI_high pd ROPE_CI ROPE_low ROPE_high
#> 1   overall -5.4817 89 -6.718247 -4.399036  1      89     -0.1       0.1
#>   ROPE_Percentage
#> 1               0
# NOT RUN {
# brms models
# -----------------------------------------------
library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
describe_posterior(model)
describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(model, ci = c(0.80, 0.90))

# BayesFactor objects
# -----------------------------------------------
library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
describe_posterior(bf)
describe_posterior(bf, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(bf, ci = c(0.80, 0.90))
# }