Describe Posterior Distributions

Compute indices relevant to describe and characterise the posterior distributions.

describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, ...)

# S3 method for numeric
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, ...)

# S3 method for stanreg
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, diagnostic = c("ESS", "Rhat"),
  priors = TRUE, effects = c("fixed", "random", "all"),
  parameters = NULL, ...)

# S3 method for brmsfit
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, diagnostic = c("ESS", "Rhat"),
  effects = c("fixed", "random", "all"), component = c("conditional",
  "zi", "zero_inflated", "all"), parameters = NULL, ...)

# S3 method for BFBayesFactor
describe_posterior(posteriors,
  centrality = "median", dispersion = FALSE, ci = 0.89,
  ci_method = "hdi", test = c("p_direction", "rope", "bf"),
  rope_range = "default", rope_ci = 0.89, priors = TRUE, ...)

Arguments

posteriors
A vector, dataframe or model of posterior draws.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Value or vector of probability of the interval (between 0 and 1) to be estimated. Named Credible Interval (CI) for consistency.

ci_method
The type of index used for Credible Interval. Can be "hdi" (default, see hdi) or "quantile" (see ci).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests.#' For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

...
Additional arguments to be passed to or from methods.

bf_prior
Distribution representing a prior for the computation of Bayes factors. Used if the input is a posterior, otherwise (in the case of models) ignored.

diagnostic
Diagnostic metrics to compute. Character (vector) or list with one or more of these options: "ESS", "Rhat", "MCSE" or "all".

priors
Add the prior used for each parameter.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Details

One or more components of point estimates (like posterior mean or median), intervals and tests can be ommitted from the summary output by setting the related argument to NULL. For example, test = NULL and centrality = NULL would only return the HDI (or CI).

References

Comparison of Point-Estimates

Region of Practical Equivalence (ROPE)

Bayes factors

Examples

library(bayestestR)

x <- rnorm(1000)
describe_posterior(x)

#>   Parameter       Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior -0.007622594 89 -1.458814 1.642115 50.3      89     -0.1       0.1
#>   ROPE_Percentage
#> 1        9.090909

describe_posterior(x, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify a prior (in the form 'prior = distribution_normal(1000, 0, 1)') to get meaningful results.

#>   Parameter       Median       MAD       Mean        SD       MAP CI    CI_low
#> 1 Posterior -0.007622594 0.9444168 0.01846216 0.9831468 -0.157443 89 -1.458814
#>    CI_high     p_map   pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage
#> 1 1.642115 0.9867397 50.3      89     -0.1       0.1        9.090909
#>   ROPE_Equivalence BF
#> 1        undecided  1

describe_posterior(x, ci = c(0.80, 0.90))

#>   Parameter       Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior -0.007622594 80 -1.156738 1.326292 50.3      89     -0.1       0.1
#> 2 Posterior -0.007622594 90 -1.502799 1.702072 50.3      89     -0.1       0.1
#>   ROPE_Percentage
#> 1        9.090909
#> 2        9.090909

df <- data.frame(replicate(4, rnorm(100)))
describe_posterior(df)

#>   Parameter      Median CI    CI_low  CI_high pd ROPE_CI ROPE_low ROPE_high
#> 1        X1 -0.04565168 89 -1.540691 1.189127 51      89     -0.1       0.1
#> 2        X2 -0.06598163 89 -1.721612 1.574241 55      89     -0.1       0.1
#> 3        X3  0.02456696 89 -1.353669 1.545773 51      89     -0.1       0.1
#> 4        X4 -0.10600216 89 -1.401507 1.113488 54      89     -0.1       0.1
#>   ROPE_Percentage
#> 1        6.666667
#> 2        8.888889
#> 3        4.444444
#> 4        6.666667

describe_posterior(df, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify priors (with column order matching 'posterior') to get meaningful results.

#>   Parameter      Median       MAD         Mean        SD          MAP CI
#> 1        X1 -0.04565168 0.9177703 -0.078794226 0.9040202  0.005049069 89
#> 2        X2 -0.06598163 1.0276300  0.017577814 1.0433150 -0.164215810 89
#> 3        X3  0.02456696 0.9886656  0.008467312 1.0260198 -0.045969999 89
#> 4        X4 -0.10600216 0.9259356 -0.056808696 0.8299671 -0.191238263 89
#>      CI_low  CI_high     p_MAP pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage
#> 1 -1.540691 1.189127 0.9999891 51      89     -0.1       0.1        6.666667
#> 2 -1.721612 1.574241 0.9804510 55      89     -0.1       0.1        8.888889
#> 3 -1.353669 1.545773 0.9990385 51      89     -0.1       0.1        4.444444
#> 4 -1.401507 1.113488 0.9883586 54      89     -0.1       0.1        6.666667
#>   ROPE_Equivalence BF
#> 1        undecided  1
#> 2        undecided  1
#> 3        undecided  1
#> 4        undecided  1

describe_posterior(df, ci = c(0.80, 0.90))

#>     Parameter      Median CI     CI_low   CI_high pd ROPE_CI ROPE_low ROPE_high
#> 1          X1 -0.04565168 80 -0.9864621 1.1477838 51      89     -0.1       0.1
#> 1.1        X1 -0.04565168 80 -0.9864621 1.1477838 51      89     -0.1       0.1
#> 3          X2 -0.06598163 80 -1.3210909 1.2621964 55      89     -0.1       0.1
#> 3.1        X2 -0.06598163 80 -1.3210909 1.2621964 55      89     -0.1       0.1
#> 5          X3  0.02456696 80 -1.3536689 1.1422376 51      89     -0.1       0.1
#> 5.1        X3  0.02456696 80 -1.3536689 1.1422376 51      89     -0.1       0.1
#> 7          X4 -0.10600216 80 -1.1722835 0.8055737 54      89     -0.1       0.1
#> 7.1        X4 -0.10600216 80 -1.1722835 0.8055737 54      89     -0.1       0.1
#>     ROPE_Percentage
#> 1          6.666667
#> 1.1        6.666667
#> 3          8.888889
#> 3.1        8.888889
#> 5          4.444444
#> 5.1        4.444444
#> 7          6.666667
#> 7.1        6.666667

# rstanarm models
# -----------------------------------------------
library(rstanarm)
model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200, refresh = 0)

#> Warning: Markov chains did not converge! Do not analyze results!

describe_posterior(model)

#>     Parameter     Median CI    CI_low    CI_high    pd ROPE_CI   ROPE_low
#> 1 (Intercept) 39.3855775 89 31.629922 47.4603765 100.0      89 -0.6026948
#> 3          wt -5.4856289 89 -6.666786 -4.4218114 100.0      89 -0.6026948
#> 2        gear -0.3046848 89 -2.133696  0.9040648  64.5      89 -0.6026948
#>   ROPE_high ROPE_Percentage ESS      Rhat Prior_Distribution Prior_Location
#> 1 0.6026948         0.00000 204 0.9957722             normal              0
#> 3 0.6026948         0.00000 215 0.9979163             normal              0
#> 2 0.6026948        49.16201 190 1.0034782             normal              0
#>   Prior_Scale
#> 1    60.26948
#> 3    15.39911
#> 2    20.42191

describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")

#> Computation of Bayes factors: sampling priors, please wait...

#>     Parameter     Median       MAD       Mean       SD        MAP CI    CI_low
#> 1 (Intercept) 39.3855775 5.6644413 39.1747056 5.544020 40.2070780 89 31.629922
#> 3          wt -5.4856289 0.7823561 -5.5232461 0.748915 -5.2672961 89 -6.666786
#> 2        gear -0.3046848 1.0004455 -0.3465336 1.007950 -0.3059815 89 -2.133696
#>      CI_high     p_MAP    pd ROPE_CI   ROPE_low ROPE_high ROPE_Percentage
#> 1 47.4603765 0.0000000 100.0      89 -0.6026948 0.6026948         0.00000
#> 3 -4.4218114 0.0000000 100.0      89 -0.6026948 0.6026948         0.00000
#> 2  0.9040648 0.9482844  64.5      89 -0.6026948 0.6026948        49.16201
#>   ROPE_Equivalence           BF ESS      Rhat Prior_Distribution Prior_Location
#> 1         rejected 7.042844e+03 204 0.9957722             normal              0
#> 3         rejected 5.551912e+04 215 0.9979163             normal              0
#> 2        undecided 4.586055e-02 190 1.0034782             normal              0
#>   Prior_Scale
#> 1    60.26948
#> 3    15.39911
#> 2    20.42191

describe_posterior(model, ci = c(0.80, 0.90))

#>       Parameter     Median CI    CI_low    CI_high    pd ROPE_CI   ROPE_low
#> 1   (Intercept) 39.3855775 80 33.003129 45.8098219 100.0      89 -0.6026948
#> 1.1 (Intercept) 39.3855775 80 33.003129 45.8098219 100.0      89 -0.6026948
#> 5            wt -5.4856289 80 -6.602624 -4.7496462 100.0      89 -0.6026948
#> 5.1          wt -5.4856289 80 -6.602624 -4.7496462 100.0      89 -0.6026948
#> 3          gear -0.3046848 80 -1.479513  0.9040648  64.5      89 -0.6026948
#> 3.1        gear -0.3046848 80 -1.479513  0.9040648  64.5      89 -0.6026948
#>     ROPE_high ROPE_Percentage ESS      Rhat Prior_Distribution Prior_Location
#> 1   0.6026948         0.00000 204 0.9957722             normal              0
#> 1.1 0.6026948         0.00000 204 0.9957722             normal              0
#> 5   0.6026948         0.00000 215 0.9979163             normal              0
#> 5.1 0.6026948         0.00000 215 0.9979163             normal              0
#> 3   0.6026948        49.16201 190 1.0034782             normal              0
#> 3.1 0.6026948        49.16201 190 1.0034782             normal              0
#>     Prior_Scale
#> 1      60.26948
#> 1.1    60.26948
#> 5      15.39911
#> 5.1    15.39911
#> 3      20.42191
#> 3.1    20.42191

# emmeans estimates
# -----------------------------------------------
library(emmeans)
describe_posterior(emtrends(model, ~1, "wt"))

#>   Parameter    Median CI    CI_low   CI_high  pd ROPE_CI ROPE_low ROPE_high
#> 1   overall -5.485629 89 -6.666786 -4.421811 100      89     -0.1       0.1
#>   ROPE_Percentage
#> 1               0

# NOT RUN {
# brms models
# -----------------------------------------------
library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
describe_posterior(model)
describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(model, ci = c(0.80, 0.90))

# BayesFactor objects
# -----------------------------------------------
library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
describe_posterior(bf)
describe_posterior(bf, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(bf, ci = c(0.80, 0.90))
# }