Describe Posterior Distributions

Compute indices relevant to describe and characterise the posterior distributions.

describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, ...)

# S3 method for numeric
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, ...)

# S3 method for stanreg
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, diagnostic = c("ESS", "Rhat"),
  priors = TRUE, effects = c("fixed", "random", "all"),
  parameters = NULL, ...)

# S3 method for brmsfit
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, diagnostic = c("ESS", "Rhat"),
  effects = c("fixed", "random", "all"), component = c("conditional",
  "zi", "zero_inflated", "all"), parameters = NULL, ...)

# S3 method for BFBayesFactor
describe_posterior(posteriors,
  centrality = "median", dispersion = FALSE, ci = 0.89,
  ci_method = "hdi", test = c("p_direction", "rope", "bf"),
  rope_range = "default", rope_ci = 0.89, priors = TRUE, ...)

Arguments

posteriors
A vector, dataframe or model of posterior draws.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "hdi" (default, see hdi) or "quantile" (see ci).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests.#' For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

...
Additional arguments to be passed to or from methods.

bf_prior
Distribution representing a prior for the computation of Bayes factors. Used if the input is a posterior, otherwise (in the case of models) ignored.

diagnostic
Diagnostic metrics to compute. Character (vector) or list with one or more of these options: "ESS", "Rhat", "MCSE" or "all".

priors
Add the prior used for each parameter.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Details

One or more components of point estimates (like posterior mean or median), intervals and tests can be ommitted from the summary output by setting the related argument to NULL. For example, test = NULL and centrality = NULL would only return the HDI (or CI).

References

Comparison of Point-Estimates

Region of Practical Equivalence (ROPE)

Bayes factors

Examples

library(bayestestR)

x <- rnorm(1000)
describe_posterior(x)

#>   Parameter       Median CI   CI_low  CI_high    pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior -0.004948617 89 -1.42606 1.655385 0.501      89     -0.1       0.1
#>   ROPE_Percentage
#> 1        7.519641

describe_posterior(x, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify a prior (in the form 'prior = distribution_normal(1000, 0, 1)') to get meaningful results.

#>   Parameter       Median       MAD        Mean        SD         MAP CI
#> 1 Posterior -0.004948617 0.9892168 0.004783671 0.9914411 -0.09815419 89
#>     CI_low  CI_high     p_map    pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage
#> 1 -1.42606 1.655385 0.9941411 0.501      89     -0.1       0.1        7.519641
#>   ROPE_Equivalence BF
#> 1        Undecided  1

describe_posterior(x, ci = c(0.80, 0.90))

#>   Parameter       Median CI    CI_low  CI_high    pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior -0.004948617 80 -1.166697 1.354956 0.501      89     -0.1       0.1
#> 2 Posterior -0.004948617 90 -1.437666 1.729997 0.501      89     -0.1       0.1
#>   ROPE_Percentage
#> 1        7.519641
#> 2        7.519641

df <- data.frame(replicate(4, rnorm(100)))
describe_posterior(df)

#>   Parameter       Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1        X1  0.009789597 89 -1.257252 2.098711 0.50      89     -0.1       0.1
#> 2        X2 -0.043368603 89 -1.068172 1.513045 0.51      89     -0.1       0.1
#> 3        X3 -0.095026444 89 -1.457298 1.890940 0.56      89     -0.1       0.1
#> 4        X4  0.142451967 89 -1.533920 1.321051 0.54      89     -0.1       0.1
#>   ROPE_Percentage
#> 1       11.111111
#> 2        5.555556
#> 3       12.222222
#> 4        7.777778

describe_posterior(df, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify priors (with column order matching 'posterior') to get meaningful results.

#>   Parameter       Median       MAD         Mean        SD           MAP CI
#> 1        X1  0.009789597 1.0035640 -0.005494185 1.0310108  0.0001242825 89
#> 2        X2 -0.043368603 0.9293306  0.104879517 0.8628621 -0.3629879765 89
#> 3        X3 -0.095026444 0.9478989 -0.058032536 1.0326693 -0.1931691221 89
#> 4        X4  0.142451967 0.9525755  0.056254365 0.9914783  0.2478383040 89
#>      CI_low  CI_high     p_MAP   pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage
#> 1 -1.257252 2.098711 0.9999998 0.50      89     -0.1       0.1       11.111111
#> 2 -1.068172 1.513045 0.9042892 0.51      89     -0.1       0.1        5.555556
#> 3 -1.457298 1.890940 0.9668645 0.56      89     -0.1       0.1       12.222222
#> 4 -1.533920 1.321051 0.9756274 0.54      89     -0.1       0.1        7.777778
#>   ROPE_Equivalence BF
#> 1        Undecided  1
#> 2        Undecided  1
#> 3        Undecided  1
#> 4        Undecided  1

describe_posterior(df, ci = c(0.80, 0.90))

#>     Parameter       Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low
#> 1          X1  0.009789597 80 -1.279227 1.172267 0.50      89     -0.1
#> 1.1        X1  0.009789597 80 -1.279227 1.172267 0.50      89     -0.1
#> 3          X2 -0.043368603 80 -1.068172 1.072403 0.51      89     -0.1
#> 3.1        X2 -0.043368603 80 -1.068172 1.072403 0.51      89     -0.1
#> 5          X3 -0.095026444 80 -1.363919 1.200763 0.56      89     -0.1
#> 5.1        X3 -0.095026444 80 -1.363919 1.200763 0.56      89     -0.1
#> 7          X4  0.142451967 80 -1.298606 1.061018 0.54      89     -0.1
#> 7.1        X4  0.142451967 80 -1.298606 1.061018 0.54      89     -0.1
#>     ROPE_high ROPE_Percentage
#> 1         0.1       11.111111
#> 1.1       0.1       11.111111
#> 3         0.1        5.555556
#> 3.1       0.1        5.555556
#> 5         0.1       12.222222
#> 5.1       0.1       12.222222
#> 7         0.1        7.777778
#> 7.1       0.1        7.777778

# rstanarm models
# -----------------------------------------------
library(rstanarm)
model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200, refresh = 0)
describe_posterior(model)

#>     Parameter     Median CI    CI_low   CI_high   pd ROPE_CI   ROPE_low
#> 1 (Intercept) 38.3557053 89 31.639685 47.166947 1.00      89 -0.6026948
#> 3          wt -5.4017938 89 -6.480299 -4.261171 1.00      89 -0.6026948
#> 2        gear -0.2358033 89 -1.623206  1.079958 0.58      89 -0.6026948
#>   ROPE_high ROPE_Percentage ESS      Rhat Prior_Distribution Prior_Location
#> 1 0.6026948         0.00000 210 1.0000646             normal              0
#> 3 0.6026948         0.00000 249 1.0055817             normal              0
#> 2 0.6026948        51.95531 206 0.9990087             normal              0
#>   Prior_Scale
#> 1    60.26948
#> 3    15.39911
#> 2    20.42191

describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")

#> Computation of Bayes factors: sampling priors, please wait...

#>     Parameter     Median       MAD       Mean        SD        MAP CI    CI_low
#> 1 (Intercept) 38.3557053 5.1770248 38.3979568 5.0427648 38.0564700 89 31.639685
#> 3          wt -5.4017938 0.7303406 -5.3984482 0.7284489 -5.3508637 89 -6.480299
#> 2        gear -0.2358033 0.9819658 -0.2600653 0.9190926 -0.1137108 89 -1.623206
#>     CI_high     p_MAP   pd ROPE_CI   ROPE_low ROPE_high ROPE_Percentage
#> 1 47.166947 0.0000000 1.00      89 -0.6026948 0.6026948         0.00000
#> 3 -4.261171 0.0000000 1.00      89 -0.6026948 0.6026948         0.00000
#> 2  1.079958 0.9941336 0.58      89 -0.6026948 0.6026948        51.95531
#>   ROPE_Equivalence           BF ESS      Rhat Prior_Distribution Prior_Location
#> 1         Rejected 1.747820e+04 210 1.0000646             normal              0
#> 3         Rejected 2.463938e+03 249 1.0055817             normal              0
#> 2        Undecided 6.149017e-02 206 0.9990087             normal              0
#>   Prior_Scale
#> 1    60.26948
#> 3    15.39911
#> 2    20.42191

describe_posterior(model, ci = c(0.80, 0.90))

#>       Parameter     Median CI    CI_low    CI_high   pd ROPE_CI   ROPE_low
#> 1   (Intercept) 38.3557053 80 32.124022 45.2015874 1.00      89 -0.6026948
#> 1.1 (Intercept) 38.3557053 80 32.124022 45.2015874 1.00      89 -0.6026948
#> 5            wt -5.4017938 80 -6.456037 -4.6871702 1.00      89 -0.6026948
#> 5.1          wt -5.4017938 80 -6.456037 -4.6871702 1.00      89 -0.6026948
#> 3          gear -0.2358033 80 -1.456527  0.7893828 0.58      89 -0.6026948
#> 3.1        gear -0.2358033 80 -1.456527  0.7893828 0.58      89 -0.6026948
#>     ROPE_high ROPE_Percentage ESS      Rhat Prior_Distribution Prior_Location
#> 1   0.6026948         0.00000 210 1.0000646             normal              0
#> 1.1 0.6026948         0.00000 210 1.0000646             normal              0
#> 5   0.6026948         0.00000 249 1.0055817             normal              0
#> 5.1 0.6026948         0.00000 249 1.0055817             normal              0
#> 3   0.6026948        51.95531 206 0.9990087             normal              0
#> 3.1 0.6026948        51.95531 206 0.9990087             normal              0
#>     Prior_Scale
#> 1      60.26948
#> 1.1    60.26948
#> 5      15.39911
#> 5.1    15.39911
#> 3      20.42191
#> 3.1    20.42191

# emmeans estimates
# -----------------------------------------------
library(emmeans)
describe_posterior(emtrends(model, ~1, "wt"))

#>   Parameter    Median CI    CI_low   CI_high pd ROPE_CI ROPE_low ROPE_high
#> 1   overall -5.401794 89 -6.480299 -4.261171  1      89     -0.1       0.1
#>   ROPE_Percentage
#> 1               0
# NOT RUN {
# brms models
# -----------------------------------------------
library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
describe_posterior(model)
describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(model, ci = c(0.80, 0.90))

# BayesFactor objects
# -----------------------------------------------
library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
describe_posterior(bf)
describe_posterior(bf, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(bf, ci = c(0.80, 0.90))
# }