Describe Posterior Distributions

Compute indices relevant to describe and characterise the posterior distributions.

describe_posterior(posteriors, centrality = "median",
  dispersion = TRUE, ci = 0.89, ci_method = "hdi", test = c("pd",
  "rope"), rope_range = "default", rope_ci = 0.89, bf_prior = NULL,
  ...)

# S3 method for stanreg
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi", test = c("pd",
  "rope"), rope_range = "default", rope_ci = 0.89, bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"), priors = TRUE, effects = c("fixed",
  "random", "all"), parameters = NULL, ...)

# S3 method for brmsfit
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi", test = c("pd",
  "rope"), rope_range = "default", rope_ci = 0.89, bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"), effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL, ...)

Arguments

posteriors
A vector, dataframe or model of posterior draws.

centrality
The point-estimate(s) (centrality indices) to compute. Can be a character or a list with "median", "mean", "MAP" or "all".

dispersion
if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Value or vector of probability of the interval (between 0 and 1) to be estimated. Named Credible Interval (CI) for consistency.

ci_method
The type of index used for Credible Interval. Can be hdi (default) or "quantile" (see ci).

test
The indices of effect existence to compute. Can be a character or a list with "p_direction", "rope", "p_map" or "bayesfactor".

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

bf_prior
Distribution representing a prior for the computation of Bayes factors. Used if the input is a posterior, otherwise (in the case of models) ignored.

...
Additional arguments to be passed to or from methods.

diagnostic
Diagnostic metrics to compute. Can be a character or a list with "ESS", "Rhat", "MCSE" or "all".

priors
Add the prior used for each parameter.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Examples

x <- rnorm(1000)
describe_posterior(x)

#>   Parameter      Median       MAD CI    CI_low  CI_high   pd ROPE_CI ROPE_low
#> 1 Posterior -0.00935952 0.9259463 89 -1.421494 1.706254 50.7      89     -0.1
#>   ROPE_high ROPE_Percentage
#> 1       0.1        10.32548

describe_posterior(x, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify a prior (in the form 'prior = distribution_normal(1000, 0, 1)') to get meaningful results.

#>   Parameter      Median       MAD       Mean        SD       MAP CI    CI_low
#> 1 Posterior -0.00935952 0.9259463 0.02165071 0.9900153 -0.139232 89 -1.421494
#>    CI_high   pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage ROPE_Equivalence BF
#> 1 1.706254 50.7      89     -0.1       0.1        10.32548        undecided  1

describe_posterior(x, ci = c(0.80, 0.90))

#>   Parameter      Median       MAD CI    CI_low  CI_high   pd ROPE_CI ROPE_low
#> 1 Posterior -0.00935952 0.9259463 80 -1.175293 1.306087 50.7      89     -0.1
#> 2 Posterior -0.00935952 0.9259463 90 -1.690634 1.587720 50.7      89     -0.1
#>   ROPE_high ROPE_Percentage
#> 1       0.1        10.32548
#> 2       0.1        10.32548

df <- data.frame(replicate(4, rnorm(100)))
describe_posterior(df)

#>   Parameter      Median       MAD CI    CI_low  CI_high pd ROPE_CI ROPE_low
#> 1        X1 -0.03752771 0.9184214 89 -1.554972 1.226449 52      89     -0.1
#> 2        X2  0.11798118 0.8793172 89 -1.691743 1.189127 54      89     -0.1
#> 3        X3  0.11025712 1.0770100 89 -1.859781 1.344945 54      89     -0.1
#> 4        X4 -0.17410893 0.8859743 89 -1.632124 1.354003 58      89     -0.1
#>   ROPE_high ROPE_Percentage
#> 1       0.1        7.777778
#> 2       0.1        6.666667
#> 3       0.1        8.888889
#> 4       0.1        7.777778

describe_posterior(df, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify priors (with columns matching 'posterior') to get meaningful results.

#>   Parameter      Median       MAD          Mean        SD          MAP CI
#> 1        X1 -0.03752771 0.9184214 -0.0990830947 0.8779604 -0.002421894 89
#> 2        X2  0.11798118 0.8793172  0.0799608218 0.9346677  0.189722256 89
#> 3        X3  0.11025712 1.0770100  0.0008557849 1.0317622  0.289246398 89
#> 4        X4 -0.17410893 0.8859743 -0.0300682134 1.0027838 -0.323350173 89
#>      CI_low  CI_high pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage
#> 1 -1.554972 1.226449 52      89     -0.1       0.1        7.777778
#> 2 -1.691743 1.189127 54      89     -0.1       0.1        6.666667
#> 3 -1.859781 1.344945 54      89     -0.1       0.1        8.888889
#> 4 -1.632124 1.354003 58      89     -0.1       0.1        7.777778
#>   ROPE_Equivalence BF
#> 1        undecided  1
#> 2        undecided  1
#> 3        undecided  1
#> 4        undecided  1

describe_posterior(df, ci = c(0.80, 0.90))

#>   Parameter      Median       MAD CI    CI_low   CI_high pd ROPE_CI ROPE_low
#> 1        X1 -0.03752771 0.9184214 80 -1.370438 0.9270811 52      89     -0.1
#> 2        X1 -0.03752771 0.9184214 90 -1.523920 1.3432494 52      89     -0.1
#> 3        X2  0.11798118 0.8793172 80 -1.108400 1.1477838 54      89     -0.1
#> 4        X2  0.11798118 0.8793172 90 -1.547880 1.4146087 54      89     -0.1
#> 5        X3  0.11025712 1.0770100 80 -1.321091 1.1412041 54      89     -0.1
#> 6        X3  0.11025712 1.0770100 90 -1.859781 1.4089197 54      89     -0.1
#> 7        X4 -0.17410893 0.8859743 80 -1.353669 1.1422376 58      89     -0.1
#> 8        X4 -0.17410893 0.8859743 90 -1.782493 1.3540026 58      89     -0.1
#>   ROPE_high ROPE_Percentage
#> 1       0.1        7.777778
#> 2       0.1        7.777778
#> 3       0.1        6.666667
#> 4       0.1        6.666667
#> 5       0.1        8.888889
#> 6       0.1        8.888889
#> 7       0.1        7.777778
#> 8       0.1        7.777778
# NOT RUN {
# rstanarm models
# -----------------------------------------------
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
describe_posterior(model)
describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(model, ci = c(0.80, 0.90))

# brms models
# -----------------------------------------------
library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
describe_posterior(model)
describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(model, ci = c(0.80, 0.90))

# BayesFactor objects
# -----------------------------------------------
library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
describe_posterior(bf)
describe_posterior(bf, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(bbf, ci = c(0.80, 0.90))
# }