Describe Posterior Distributions

Compute indices relevant to describe and characterise the posterior distributions.

describe_posterior(posteriors, estimate = "median", ci = 0.9,
  ci_method = "hdi", test = c("pd", "rope"), rope_range = "default",
  rope_full = TRUE, dispersion = TRUE, ...)

# S3 method for stanreg
describe_posterior(posteriors, estimate = "median",
  ci = 0.9, ci_method = "hdi", test = c("pd", "rope"),
  rope_range = "default", rope_full = TRUE, dispersion = TRUE,
  effects = c("fixed", "random", "all"), parameters = NULL, ...)

# S3 method for brmsfit
describe_posterior(posteriors, estimate = "median",
  ci = 0.9, ci_method = "hdi", test = c("pd", "rope"),
  rope_range = "default", rope_full = TRUE, dispersion = TRUE,
  effects = c("fixed", "random", "all"), component = c("conditional",
  "zi", "zero_inflated", "all"), parameters = NULL, ...)

Arguments

posteriors
A vector, dataframe or model of posterior draws.

estimate
The point-estimate(s) to compute. Can be a character or a list with "median", "mean" or "MAP".

ci
Value or vector of probability of the interval (between 0 and 1) to be estimated. Named Credible Interval (CI) for consistency.

ci_method
The type of index used for Credible Interval. Can be hdi (default) or "quantile" (see ci).

test
The indices of effect existence to compute. Can be a character or a list with "p_direction", "rope", "p_map" or "bayesfactor".

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_full
If TRUE, use the proportion of the entire posterior distribution for the equivalence test. Otherwise, use the proportion of HDI as indicated by the ci argument.

dispersion
if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

...
Additional arguments to be passed to or from methods.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Examples

describe_posterior(rnorm(1000))

#>         Median       MAD    CI_low  CI_high   pd ROPE_Percentage
#> 1 -0.005473077 0.9636613 -1.405601 1.712907 50.2             8.5
#>   ROPE_Equivalence
#> 1        undecided

df <- data.frame(replicate(4, rnorm(100)))
describe_posterior(df, estimate = "all", dispersion = TRUE)

#>   Parameter      Median       MAD        Mean        SD          MAP    CI_low
#> 1        X1  0.10526723 1.0253289  0.05899111 1.0194272  0.001657595 -1.689547
#> 2        X2  0.06909295 1.0267032  0.03713839 0.9703200 -0.023262587 -1.365789
#> 3        X3  0.01312073 0.9436938  0.03503264 0.9869062 -0.115885664 -1.681573
#> 4        X4 -0.09898537 1.1543220 -0.06861218 1.0414355 -0.108738311 -1.587608
#>    CI_high pd ROPE_Percentage ROPE_Equivalence
#> 1 1.510348 52               4        undecided
#> 2 1.758286 53               9        undecided
#> 3 1.539376 50               9        undecided
#> 4 1.526738 53               9        undecided

describe_posterior(df, estimate = c("median", "MAP"))

#>   Parameter      Median       MAD          MAP    CI_low  CI_high pd
#> 1        X1  0.10526723 1.0253289  0.001657595 -1.689547 1.510348 52
#> 2        X2  0.06909295 1.0267032 -0.023262587 -1.365789 1.758286 53
#> 3        X3  0.01312073 0.9436938 -0.115885664 -1.681573 1.539376 50
#> 4        X4 -0.09898537 1.1543220 -0.108738311 -1.587608 1.526738 53
#>   ROPE_Percentage ROPE_Equivalence
#> 1               4        undecided
#> 2               9        undecided
#> 3               9        undecided
#> 4               9        undecided