Describe Posterior Distributions

Compute indices relevant to describe and characterise the posterior distributions.

describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  ...
)

# S3 method for numeric
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  bf_prior = NULL,
  BF = 1,
  ...
)

# S3 method for stanreg
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"),
  priors = FALSE,
  effects = c("fixed", "random", "all"),
  parameters = NULL,
  BF = 1,
  ...
)

# S3 method for stanmvreg
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = "p_direction",
  rope_range = "default",
  rope_ci = 0.89,
  bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"),
  priors = FALSE,
  effects = c("fixed", "random", "all"),
  parameters = NULL,
  ...
)

# S3 method for MCMCglmm
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  diagnostic = "ESS",
  parameters = NULL,
  ...
)

# S3 method for brmsfit
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope"),
  rope_range = "default",
  rope_ci = 0.89,
  bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"),
  effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL,
  BF = 1,
  ...
)

# S3 method for BFBayesFactor
describe_posterior(
  posteriors,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = c("p_direction", "rope", "bf"),
  rope_range = "default",
  rope_ci = 0.89,
  priors = TRUE,
  ...
)

Arguments

posteriors
A vector, dataframe or model of posterior draws.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi), "ETI" (see eti) or "SI" (see si).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

...
Additional arguments to be passed to or from methods.

bf_prior
Distribution representing a prior for the computation of Bayes factors / SI. Used if the input is a posterior, otherwise (in the case of models) ignored.

BF
The amount of support required to be included in the support interval.

diagnostic
Diagnostic metrics to compute. Character (vector) or list with one or more of these options: "ESS", "Rhat", "MCSE" or "all".

priors
Add the prior used for each parameter.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Details

One or more components of point estimates (like posterior mean or median), intervals and tests can be ommitted from the summary output by setting the related argument to NULL. For example, test = NULL and centrality = NULL would only return the HDI (or CI).

References

Comparison of Point-Estimates

Region of Practical Equivalence (ROPE)

Bayes factors

Examples

library(bayestestR)

x <- rnorm(1000)
describe_posterior(x)

#>   Parameter      Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior -0.07114759 89 -1.585473 1.460772 0.53      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.08754209

describe_posterior(x, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify a prior (in the form 'prior = distribution_normal(1000, 0, 1)') to get meaningful results.

#>   Parameter      Median      MAD        Mean        SD        MAP CI    CI_low
#> 1 Posterior -0.07114759 1.022044 -0.03122015 0.9749539 -0.2118201 89 -1.585473
#>    CI_high     p_map   pd ROPE_low.x ROPE_high.x p_ROPE    ps ROPE_CI
#> 1 1.460772 0.9623947 0.53       -0.1         0.1  0.078 0.492      89
#>   ROPE_low.y ROPE_high.y ROPE_Percentage ROPE_Equivalence BF
#> 1       -0.1         0.1      0.08754209        Undecided  1

describe_posterior(x, ci = c(0.80, 0.90))

#>   Parameter      Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior -0.07114759 80 -1.238857 1.189950 0.53      89     -0.1       0.1
#> 2 Posterior -0.07114759 90 -1.665585 1.449041 0.53      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.08754209
#> 2      0.08754209

df <- data.frame(replicate(4, rnorm(100)))
describe_posterior(df)

#>   Parameter      Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1        X1 -0.14501191 89 -1.687645 1.614477 0.53      89     -0.1       0.1
#> 2        X2  0.16014275 89 -1.723535 1.682682 0.52      89     -0.1       0.1
#> 3        X3 -0.04019892 89 -2.116266 1.241637 0.52      89     -0.1       0.1
#> 4        X4  0.20234268 89 -2.062503 1.417986 0.56      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.05555556
#> 2      0.03333333
#> 3      0.13333333
#> 4      0.10000000

describe_posterior(df, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify priors (with column order matching 'posterior') to get meaningful results.

#>   Parameter      Median       MAD         Mean       SD        MAP CI    CI_low
#> 1        X1 -0.14501191 1.0990740 -0.031749567 1.107530 -0.3672558 89 -1.687645
#> 2        X2  0.16014275 1.2363190  0.004776539 1.147462  0.2809465 89 -1.723535
#> 3        X3 -0.04019892 1.0889390 -0.142847925 1.089237  0.1507107 89 -2.116266
#> 4        X4  0.20234268 0.9078906  0.216717183 1.050562  0.1233038 89 -2.062503
#>    CI_high     p_MAP   pd ROPE_low.x ROPE_high.x p_ROPE   ps ROPE_CI ROPE_low.y
#> 1 1.614477 0.9286150 0.53       -0.1         0.1   0.05 0.51      89       -0.1
#> 2 1.682682 0.9798509 0.52       -0.1         0.1   0.03 0.51      89       -0.1
#> 3 1.241637 0.9926612 0.52       -0.1         0.1   0.12 0.46      89       -0.1
#> 4 1.417986 0.9878487 0.56       -0.1         0.1   0.09 0.54      89       -0.1
#>   ROPE_high.y ROPE_Percentage ROPE_Equivalence BF
#> 1         0.1      0.05555556        Undecided  1
#> 2         0.1      0.03333333        Undecided  1
#> 3         0.1      0.13333333        Undecided  1
#> 4         0.1      0.10000000        Undecided  1

describe_posterior(df, ci = c(0.80, 0.90))

#>   Parameter      Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1        X1 -0.14501191 80 -1.330716 1.298955 0.53      89     -0.1       0.1
#> 2        X1 -0.14501191 90 -1.687645 1.770111 0.53      89     -0.1       0.1
#> 3        X2  0.16014275 80 -1.363448 1.446004 0.52      89     -0.1       0.1
#> 4        X2  0.16014275 90 -1.723535 1.737621 0.52      89     -0.1       0.1
#> 5        X3 -0.04019892 80 -1.181482 1.590790 0.52      89     -0.1       0.1
#> 6        X3 -0.04019892 90 -1.956193 1.469050 0.52      89     -0.1       0.1
#> 7        X4  0.20234268 80 -1.139573 1.206828 0.56      89     -0.1       0.1
#> 8        X4  0.20234268 90 -1.291709 2.374397 0.56      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.05555556
#> 2      0.05555556
#> 3      0.03333333
#> 4      0.03333333
#> 5      0.13333333
#> 6      0.13333333
#> 7      0.10000000
#> 8      0.10000000

# rstanarm models
# -----------------------------------------------
library(rstanarm)
model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200, refresh = 0)

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

describe_posterior(model)

#> # Description of Posterior Distributions
#> 
#>    Parameter Median CI CI_low CI_high   pd ROPE_CI ROPE_low ROPE_high
#>  (Intercept) 39.673 89  30.85  47.424 1.00      89   -0.603     0.603
#>           wt -5.559 89  -6.76  -4.442 1.00      89   -0.603     0.603
#>         gear -0.448 89  -1.98   0.891 0.69      89   -0.603     0.603
#>  ROPE_Percentage  Rhat ESS
#>            0.000 0.996 220
#>            0.000 0.996 200
#>            0.503 0.993 215
#> 

describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")

#> Computation of Bayes factors: sampling priors, please wait...

#> # Description of Posterior Distributions
#> 
#>  Fixed Effects (Conditional Model)
#> 
#>    Parameter Median   MAD   Mean    SD    MAP CI CI_low CI_high p_MAP   pd
#>  (Intercept) 39.673 5.228 39.441 5.290 41.407 89  30.85  47.424 0.000 1.00
#>           wt -5.559 0.684 -5.562 0.708 -5.627 89  -6.76  -4.442 0.000 1.00
#>         gear -0.448 1.018 -0.441 0.954 -0.202 89  -1.98   0.891 0.958 0.69
#>  ROPE_low.x ROPE_high.x p_ROPE    ps ROPE_CI ROPE_low.y ROPE_high.y
#>          NA          NA     NA 1.000      89     -0.603       0.603
#>          NA          NA     NA 1.000      89     -0.603       0.603
#>          NA          NA     NA 0.425      89     -0.603       0.603
#>  ROPE_Percentage ROPE_Equivalence       BF  Rhat ESS
#>            0.000         Rejected 3.82e+03 0.996 220
#>            0.000         Rejected 1.77e+04 0.996 200
#>            0.503        Undecided 4.03e-02 0.993 215
#> 
#> 
#> 
#>    Parameter Median MAD Mean SD MAP CI CI_low CI_high p_MAP pd ROPE_low.x
#>  (Intercept)     NA  NA   NA NA  NA NA     NA      NA    NA NA     -0.603
#>         gear     NA  NA   NA NA  NA NA     NA      NA    NA NA     -0.603
#>           wt     NA  NA   NA NA  NA NA     NA      NA    NA NA     -0.603
#>  ROPE_high.x p_ROPE ps ROPE_CI ROPE_low.y ROPE_high.y ROPE_Percentage
#>        0.603   0.00 NA      NA         NA          NA              NA
#>        0.603   0.45 NA      NA         NA          NA              NA
#>        0.603   0.00 NA      NA         NA          NA              NA
#>  ROPE_Equivalence BF  Rhat ESS
#>              <NA> NA 0.996 220
#>              <NA> NA 0.993 215
#>              <NA> NA 0.996 200
#> 

describe_posterior(model, ci = c(0.80, 0.90))

#> # Description of Posterior Distributions
#> 
#>    Parameter Median CI CI_low CI_high   pd ROPE_CI ROPE_low ROPE_high
#>  (Intercept) 39.673 80  32.40  45.708 1.00      89   -0.603     0.603
#>  (Intercept) 39.673 90  31.76  48.705 1.00      89   -0.603     0.603
#>           wt -5.559 80  -6.43  -4.638 1.00      89   -0.603     0.603
#>           wt -5.559 90  -6.67  -4.281 1.00      89   -0.603     0.603
#>         gear -0.448 80  -1.78   0.577 0.69      89   -0.603     0.603
#>         gear -0.448 90  -1.98   1.028 0.69      89   -0.603     0.603
#>  ROPE_Percentage  Rhat ESS
#>            0.000 0.996 220
#>            0.000 0.996 220
#>            0.000 0.996 200
#>            0.000 0.996 200
#>            0.503 0.993 215
#>            0.503 0.993 215
#> 

# emmeans estimates
# -----------------------------------------------
library(emmeans)

#> Welcome to emmeans.
#> NOTE -- Important change from versions <= 1.41:
#>     Indicator predictors are now treated as 2-level factors by default.
#>     To revert to old behavior, use emm_options(cov.keep = character(0))

describe_posterior(emtrends(model, ~1, "wt"))

#> Warning: restarting interrupted promise evaluation

#> Error in as.vector(x) : object 'model' not found

#> Error in ref_grid(object = model, options = list(just.data = TRUE)): Perhaps a 'data' or 'params' argument is needed

if (FALSE) {
# brms models
# -----------------------------------------------
library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
describe_posterior(model)
describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(model, ci = c(0.80, 0.90))

# BayesFactor objects
# -----------------------------------------------
library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
describe_posterior(bf)
describe_posterior(bf, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(bf, ci = c(0.80, 0.90))
}