Describe Posterior Distributions

Compute indices relevant to describe and characterise the posterior distributions.

describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, ...)

# S3 method for numeric
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, ...)

# S3 method for stanreg
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, diagnostic = c("ESS", "Rhat"),
  priors = FALSE, effects = c("fixed", "random", "all"),
  parameters = NULL, ...)

# S3 method for MCMCglmm
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, diagnostic = "ESS", parameters = NULL, ...)

# S3 method for brmsfit
describe_posterior(posteriors, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi",
  test = c("p_direction", "rope"), rope_range = "default",
  rope_ci = 0.89, bf_prior = NULL, diagnostic = c("ESS", "Rhat"),
  effects = c("fixed", "random", "all"), component = c("conditional",
  "zi", "zero_inflated", "all"), parameters = NULL, ...)

# S3 method for BFBayesFactor
describe_posterior(posteriors,
  centrality = "median", dispersion = FALSE, ci = 0.89,
  ci_method = "hdi", test = c("p_direction", "rope", "bf"),
  rope_range = "default", rope_ci = 0.89, priors = TRUE, ...)

Arguments

posteriors
A vector, dataframe or model of posterior draws.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi) or "ETI" (see eti).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

...
Additional arguments to be passed to or from methods.

bf_prior
Distribution representing a prior for the computation of Bayes factors. Used if the input is a posterior, otherwise (in the case of models) ignored.

diagnostic
Diagnostic metrics to compute. Character (vector) or list with one or more of these options: "ESS", "Rhat", "MCSE" or "all".

priors
Add the prior used for each parameter.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Details

One or more components of point estimates (like posterior mean or median), intervals and tests can be ommitted from the summary output by setting the related argument to NULL. For example, test = NULL and centrality = NULL would only return the HDI (or CI).

References

Comparison of Point-Estimates

Region of Practical Equivalence (ROPE)

Bayes factors

Examples

library(bayestestR)

x <- rnorm(1000)
describe_posterior(x)

#>   Parameter       Median CI    CI_low  CI_high    pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior -0.002118877 89 -1.591667 1.523491 0.502      89     -0.1       0.1
#>   ROPE_Percentage
#> 1       0.1021324

describe_posterior(x, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify a prior (in the form 'prior = distribution_normal(1000, 0, 1)') to get meaningful results.

#>   Parameter       Median       MAD         Mean        SD        MAP CI
#> 1 Posterior -0.002118877 0.9223343 -0.006218011 0.9717942 -0.0298333 89
#>      CI_low  CI_high     p_map    pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage
#> 1 -1.591667 1.523491 0.9990962 0.502      89     -0.1       0.1       0.1021324
#>   ROPE_Equivalence BF
#> 1        Undecided  1

describe_posterior(x, ci = c(0.80, 0.90))

#>   Parameter       Median CI    CI_low  CI_high    pd ROPE_CI ROPE_low ROPE_high
#> 1 Posterior -0.002118877 80 -1.285348 1.126897 0.502      89     -0.1       0.1
#> 2 Posterior -0.002118877 90 -1.553458 1.699002 0.502      89     -0.1       0.1
#>   ROPE_Percentage
#> 1       0.1021324
#> 2       0.1021324

df <- data.frame(replicate(4, rnorm(100)))
describe_posterior(df)

#>   Parameter      Median CI    CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1        X1 -0.08229185 89 -1.676301 1.847630 0.56      89     -0.1       0.1
#> 2        X2  0.04976194 89 -1.269663 1.496096 0.54      89     -0.1       0.1
#> 3        X3  0.12755187 89 -2.028607 1.361688 0.57      89     -0.1       0.1
#> 4        X4  0.30169076 89 -1.317348 2.209598 0.60      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.08888889
#> 2      0.08888889
#> 3      0.07777778
#> 4      0.10000000

describe_posterior(df, centrality = "all", dispersion = TRUE, test = "all")

#> Warning: Prior not specified! Please specify priors (with column order matching 'posterior') to get meaningful results.

#>   Parameter      Median       MAD        Mean        SD        MAP CI    CI_low
#> 1        X1 -0.08229185 1.4797008 -0.15164738 1.2112597 0.88861187 89 -1.676301
#> 2        X2  0.04976194 0.9263314  0.09558829 0.9455221 0.06664366 89 -1.269663
#> 3        X3  0.12755187 0.9416361  0.05871179 1.0530845 0.33459139 89 -2.028607
#> 4        X4  0.30169076 0.9697433  0.24824422 1.0925816 0.28901625 89 -1.317348
#>    CI_high     p_MAP   pd ROPE_CI ROPE_low ROPE_high ROPE_Percentage
#> 1 1.847630 0.8648198 0.56      89     -0.1       0.1      0.08888889
#> 2 1.496096 0.9972080 0.54      89     -0.1       0.1      0.08888889
#> 3 1.361688 0.9325454 0.57      89     -0.1       0.1      0.07777778
#> 4 2.209598 0.9761697 0.60      89     -0.1       0.1      0.10000000
#>   ROPE_Equivalence BF
#> 1        Undecided  1
#> 2        Undecided  1
#> 3        Undecided  1
#> 4        Undecided  1

describe_posterior(df, ci = c(0.80, 0.90))

#>   Parameter      Median CI     CI_low  CI_high   pd ROPE_CI ROPE_low ROPE_high
#> 1        X1 -0.08229185 80 -1.6763008 1.184397 0.56      89     -0.1       0.1
#> 2        X1 -0.08229185 90 -1.6971130 1.847630 0.56      89     -0.1       0.1
#> 3        X2  0.04976194 80 -1.1686963 1.085133 0.54      89     -0.1       0.1
#> 4        X2  0.04976194 90 -1.3642485 1.496096 0.54      89     -0.1       0.1
#> 5        X3  0.12755187 80 -1.3150526 1.220136 0.57      89     -0.1       0.1
#> 6        X3  0.12755187 90 -2.0286072 1.456001 0.57      89     -0.1       0.1
#> 7        X4  0.30169076 80 -0.8802615 1.514967 0.60      89     -0.1       0.1
#> 8        X4  0.30169076 90 -1.3173483 2.401810 0.60      89     -0.1       0.1
#>   ROPE_Percentage
#> 1      0.08888889
#> 2      0.08888889
#> 3      0.08888889
#> 4      0.08888889
#> 5      0.07777778
#> 6      0.07777778
#> 7      0.10000000
#> 8      0.10000000

# rstanarm models
# -----------------------------------------------
library(rstanarm)
model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200, refresh = 0)

#> Warning: The largest R-hat is 1.11, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: Markov chains did not converge! Do not analyze results!

describe_posterior(model)

#> # Description of Posterior Distributions
#> 
#>    Parameter Median CI CI_low CI_high    pd ROPE_CI ROPE_low ROPE_high
#>  (Intercept) 39.618 89  31.37  47.392 1.000      89   -0.603     0.603
#>           wt -5.493 89  -6.84  -4.590 1.000      89   -0.603     0.603
#>         gear -0.347 89  -1.79   0.955 0.635      89   -0.603     0.603
#>  ROPE_Percentage  Rhat ESS
#>            0.000 0.997 216
#>            0.000 0.993 158
#>            0.458 0.992 219
#> 

describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")

#> Computation of Bayes factors: sampling priors, please wait...

#> # Description of Posterior Distributions
#> 
#>    Parameter Median   MAD   Mean    SD    MAP CI CI_low CI_high p_MAP    pd
#>  (Intercept) 39.618 5.096 39.340 5.083 39.929 89  31.37  47.392 0.000 1.000
#>           wt -5.493 0.694 -5.535 0.699 -5.176 89  -6.84  -4.590 0.000 1.000
#>         gear -0.347 1.111 -0.363 0.929 -0.369 89  -1.79   0.955 0.952 0.635
#>  ROPE_CI ROPE_low ROPE_high ROPE_Percentage ROPE_Equivalence       BF  Rhat ESS
#>       89   -0.603     0.603           0.000         Rejected 4.75e+04 0.997 216
#>       89   -0.603     0.603           0.000         Rejected 7.64e+14 0.993 158
#>       89   -0.603     0.603           0.458        Undecided 4.84e-02 0.992 219
#> 

describe_posterior(model, ci = c(0.80, 0.90))

#> # Description of Posterior Distributions
#> 
#>    Parameter Median CI CI_low CI_high    pd ROPE_CI ROPE_low ROPE_high
#>  (Intercept) 39.618 80  32.04  44.932 1.000      89   -0.603     0.603
#>  (Intercept) 39.618 90  31.27  47.612 1.000      89   -0.603     0.603
#>           wt -5.493 80  -6.50  -4.797 1.000      89   -0.603     0.603
#>           wt -5.493 90  -6.84  -4.494 1.000      89   -0.603     0.603
#>         gear -0.347 80  -1.54   0.782 0.635      89   -0.603     0.603
#>         gear -0.347 90  -1.96   0.873 0.635      89   -0.603     0.603
#>  ROPE_Percentage  Rhat ESS
#>            0.000 0.997 216
#>            0.000 0.997 216
#>            0.000 0.993 158
#>            0.000 0.993 158
#>            0.458 0.992 219
#>            0.458 0.992 219
#> 

# emmeans estimates
# -----------------------------------------------
library(emmeans)

#> Warning: package 'emmeans' was built under R version 3.6.1

describe_posterior(emtrends(model, ~1, "wt"))

#>   Parameter    Median CI    CI_low   CI_high pd ROPE_CI ROPE_low ROPE_high
#> 1   overall -5.493422 89 -6.842873 -4.589787  1      89     -0.1       0.1
#>   ROPE_Percentage
#> 1               0

if (FALSE) {
# brms models
# -----------------------------------------------
library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
describe_posterior(model)
describe_posterior(model, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(model, ci = c(0.80, 0.90))

# BayesFactor objects
# -----------------------------------------------
library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
describe_posterior(bf)
describe_posterior(bf, centrality = "all", dispersion = TRUE, test = "all")
describe_posterior(bf, ci = c(0.80, 0.90))
}