Diagnostic values for each iteration

Returns the accumulated log-posterior, the average Metropolis acceptance rate, divergent transitions, treedepth rather than terminated its evolution normally.

diagnostic_draws(posteriors, ...)

Arguments

posteriors
A stanreg or brms model.

...
Currently not used.

Examples

# \donttest{
set.seed(333)

if (require("brms", quietly = TRUE)) {
  model <- brm(mpg ~ wt * cyl * vs,
    data = mtcars,
    iter = 100, control = list(adapt_delta = 0.80),
    refresh = 0
  )
  diagnostic_draws(model)
}

#> Loading 'brms' package (version 2.15.0). Useful instructions
#> can be found by typing help('brms'). A more detailed introduction
#> to the package is available through vignette('brms_overview').

#> 
#> Attaching package: ‘brms’

#> The following objects are masked from ‘package:rstanarm’:
#> 
#>     dirichlet, exponential, get_y, lasso, ngrps

#> The following object is masked from ‘package:stats’:
#> 
#>     ar

#> Compiling Stan program...

#> Start sampling

#> Warning: There were 65 transitions after warmup that exceeded the maximum treedepth. Increase max_treedepth above 10. See
#> http://mc-stan.org/misc/warnings.html#maximum-treedepth-exceeded

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: The largest R-hat is 1.88, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#>     Chain Iteration Acceptance_Rate  Step_Size Tree_Depth n_Leapfrog Divergent
#> 1       1         1       0.9718656 0.03930745          8        383         0
#> 2       1         2       0.9179547 0.03930745         10       1023         0
#> 3       1         3       0.9996224 0.03930745         10       1023         0
#> 4       1         4       0.9948554 0.03930745          8        255         0
#> 5       1         5       0.8665508 0.03930745          9        767         0
#> 6       1         6       0.8327168 0.03930745          9        575         0
#> 7       1         7       0.9620037 0.03930745         10       1023         0
#> 8       1         8       0.9679693 0.03930745         10       1023         0
#> 9       1         9       0.7740241 0.03930745          9        895         0
#> 10      1        10       0.9823040 0.03930745          6        127         0
#> 11      1        11       0.9697384 0.03930745         10       1023         0
#> 12      1        12       0.9327940 0.03930745          6         95         0
#> 13      1        13       0.9987514 0.03930745          6        127         0
#> 14      1        14       0.9539308 0.03930745          6        127         0
#> 15      1        15       0.9542032 0.03930745         10       1023         0
#> 16      1        16       0.7951302 0.03930745         10       1023         0
#> 17      1        17       0.9849405 0.03930745          6        127         0
#> 18      1        18       0.9994078 0.03930745          7        127         0
#> 19      1        19       0.9966464 0.03930745          6        127         0
#> 20      1        20       0.9363729 0.03930745          8        511         0
#> 21      1        21       0.9478299 0.03930745          6         63         0
#> 22      1        22       0.6934296 0.03930745         10       1023         0
#> 23      1        23       0.8472706 0.03930745          6        127         0
#> 24      1        24       0.9882368 0.03930745          9        639         0
#> 25      1        25       0.9111610 0.03930745          5         63         0
#> 26      1        26       0.9076800 0.03930745         10       1023         0
#> 27      1        27       0.4333427 0.03930745         10       1023         0
#> 28      1        28       0.9672924 0.03930745          9        575         0
#> 29      1        29       0.9273097 0.03930745         10       1023         0
#> 30      1        30       0.9662291 0.03930745         10       1023         0
#> 31      1        31       0.9936620 0.03930745         10       1023         0
#> 32      1        32       0.9848061 0.03930745          6        127         0
#> 33      1        33       0.9970402 0.03930745          6        127         0
#> 34      1        34       0.9313294 0.03930745          7        255         0
#> 35      1        35       0.9745841 0.03930745          8        511         0
#> 36      1        36       0.7751155 0.03930745          7        159         0
#> 37      1        37       0.9971050 0.03930745         10       1023         0
#> 38      1        38       0.9927280 0.03930745         10       1023         0
#> 39      1        39       0.9461566 0.03930745          6        127         0
#> 40      1        40       0.7932146 0.03930745          9        575         0
#> 41      1        41       0.8903876 0.03930745          4         31         0
#> 42      1        42       0.9675546 0.03930745          5         47         0
#> 43      1        43       0.9976433 0.03930745          6        127         0
#> 44      1        44       0.9822572 0.03930745          5         31         0
#> 45      1        45       0.8385623 0.03930745          6         63         0
#> 46      1        46       0.9279874 0.03930745          7        191         0
#> 47      1        47       0.9980570 0.03930745         10       1023         0
#> 48      1        48       0.7608691 0.03930745          9        767         0
#> 49      1        49       0.8870656 0.03930745          5         31         0
#> 50      1        50       0.9968085 0.03930745          9        575         0
#> 51      2         1       0.9772609 0.03609483          7        255         0
#> 52      2         2       0.6922982 0.03609483         10       1023         0
#> 53      2         3       0.9627171 0.03609483          8        383         0
#> 54      2         4       0.9420202 0.03609483          6        127         0
#> 55      2         5       0.9945715 0.03609483          9        639         0
#> 56      2         6       0.9296026 0.03609483         10       1023         0
#> 57      2         7       0.9998052 0.03609483         10       1023         0
#> 58      2         8       0.7433905 0.03609483         10       1023         0
#> 59      2         9       0.8607367 0.03609483          8        511         0
#> 60      2        10       0.6414925 0.03609483          9        895         0
#> 61      2        11       0.9906558 0.03609483         10       1023         0
#> 62      2        12       0.9482264 0.03609483         10       1023         0
#> 63      2        13       0.9229002 0.03609483          6        127         0
#> 64      2        14       0.9974470 0.03609483         10       1023         0
#> 65      2        15       0.9828063 0.03609483         10       1023         0
#> 66      2        16       0.9456110 0.03609483          5         63         0
#> 67      2        17       0.9100028 0.03609483          9        831         0
#> 68      2        18       1.0000000 0.03609483          8        383         0
#> 69      2        19       0.9826515 0.03609483         10       1023         0
#> 70      2        20       0.9985309 0.03609483          6         63         0
#> 71      2        21       0.9538399 0.03609483          6         63         0
#> 72      2        22       0.8648729 0.03609483          6         63         0
#> 73      2        23       0.9929837 0.03609483          9        767         0
#> 74      2        24       0.9887161 0.03609483          6         63         0
#> 75      2        25       0.9531217 0.03609483          6        127         0
#> 76      2        26       0.8358533 0.03609483         10       1023         0
#> 77      2        27       0.9031216 0.03609483         10       1023         0
#> 78      2        28       0.8661320 0.03609483          8        447         0
#> 79      2        29       0.9515822 0.03609483         10       1023         0
#> 80      2        30       0.9956062 0.03609483         10       1023         0
#> 81      2        31       0.9996696 0.03609483          6         63         0
#> 82      2        32       0.8865861 0.03609483          7        127         0
#> 83      2        33       0.9974405 0.03609483         10       1023         0
#> 84      2        34       0.9261094 0.03609483          7        191         0
#> 85      2        35       0.9536453 0.03609483          7        247         0
#> 86      2        36       0.9820820 0.03609483          8        511         0
#> 87      2        37       0.9691326 0.03609483         10       1023         0
#> 88      2        38       0.8470522 0.03609483          7        127         0
#> 89      2        39       0.7748773 0.03609483          7        127         0
#> 90      2        40       0.9901053 0.03609483          9       1023         0
#> 91      2        41       0.9282376 0.03609483          7        255         0
#> 92      2        42       0.9788186 0.03609483          6        103         0
#> 93      2        43       0.9973325 0.03609483         10       1023         0
#> 94      2        44       0.8166014 0.03609483          9        879         0
#> 95      2        45       0.9994634 0.03609483         10       1023         0
#> 96      2        46       0.9770366 0.03609483          9        639         0
#> 97      2        47       0.9958370 0.03609483         10       1023         0
#> 98      2        48       0.9980731 0.03609483         10       1023         0
#> 99      2        49       0.9313372 0.03609483         10       1023         0
#> 100     2        50       0.9997957 0.03609483          9        959         0
#> 101     3         1       0.9105773 0.03535163         10       1023         0
#> 102     3         2       0.8846014 0.03535163          9        895         0
#> 103     3         3       0.9932689 0.03535163          6        127         0
#> 104     3         4       0.8878102 0.03535163          6         63         0
#> 105     3         5       0.9985033 0.03535163          7        255         0
#> 106     3         6       0.9406700 0.03535163          7        191         0
#> 107     3         7       0.8087154 0.03535163          7        159         0
#> 108     3         8       0.9972134 0.03535163         10       1023         0
#> 109     3         9       0.9557490 0.03535163          9        767         0
#> 110     3        10       0.5339378 0.03535163         10       1023         0
#> 111     3        11       1.0000000 0.03535163          5         31         0
#> 112     3        12       0.9799325 0.03535163          6         63         0
#> 113     3        13       0.9952303 0.03535163         10       1023         0
#> 114     3        14       0.9503720 0.03535163          6         63         0
#> 115     3        15       0.9979172 0.03535163          6         63         0
#> 116     3        16       0.6295366 0.03535163          5         47         0
#> 117     3        17       0.8145478 0.03535163          4         31         0
#> 118     3        18       0.9118755 0.03535163          8        439         0
#> 119     3        19       0.8306728 0.03535163          9       1023         0
#> 120     3        20       0.8451056 0.03535163          6         63         0
#> 121     3        21       0.9878328 0.03535163          7        191         0
#> 122     3        22       0.9392717 0.03535163         10       1023         0
#> 123     3        23       0.9727489 0.03535163         10       1023         0
#> 124     3        24       0.9777867 0.03535163          9        575         0
#> 125     3        25       0.9928800 0.03535163         10       1023         0
#> 126     3        26       0.9950480 0.03535163          8        511         0
#> 127     3        27       0.9915503 0.03535163          6         63         0
#> 128     3        28       0.8113916 0.03535163          8        511         0
#> 129     3        29       0.9587285 0.03535163          7        255         0
#> 130     3        30       0.8374815 0.03535163          9        959         0
#> 131     3        31       0.9711685 0.03535163         10       1023         0
#> 132     3        32       0.9943325 0.03535163         10       1023         0
#> 133     3        33       0.9833118 0.03535163         10       1023         0
#> 134     3        34       0.9612838 0.03535163          8        271         0
#> 135     3        35       0.9952010 0.03535163          6        127         0
#> 136     3        36       0.9986718 0.03535163          7        191         0
#> 137     3        37       0.9389936 0.03535163         10       1023         0
#> 138     3        38       0.9858870 0.03535163          7        255         0
#> 139     3        39       0.9850289 0.03535163         10       1023         0
#> 140     3        40       0.9670281 0.03535163          8        511         0
#> 141     3        41       0.9833927 0.03535163          6         95         0
#> 142     3        42       0.7629058 0.03535163          7        127         0
#> 143     3        43       0.9591085 0.03535163          7        127         0
#> 144     3        44       0.9665914 0.03535163          7        127         0
#> 145     3        45       0.8986125 0.03535163          6         63         0
#> 146     3        46       0.9938699 0.03535163         10       1023         0
#> 147     3        47       0.9274898 0.03535163          6        127         0
#> 148     3        48       0.9223965 0.03535163          9        735         0
#> 149     3        49       0.9927304 0.03535163         10       1023         0
#> 150     3        50       0.9305499 0.03535163          8        319         0
#> 151     4         1       0.9986462 0.03198024         10       1023         0
#> 152     4         2       0.9552077 0.03198024         10       1023         0
#> 153     4         3       0.9506264 0.03198024         10       1023         0
#> 154     4         4       0.9886305 0.03198024         10       1023         0
#> 155     4         5       0.9965268 0.03198024          7        191         0
#> 156     4         6       0.8855937 0.03198024          9        743         0
#> 157     4         7       0.9951101 0.03198024          9        767         0
#> 158     4         8       0.9448054 0.03198024          7        191         0
#> 159     4         9       0.9701187 0.03198024          6         63         0
#> 160     4        10       0.9845182 0.03198024          8        511         0
#> 161     4        11       0.9995760 0.03198024         10       1023         0
#> 162     4        12       0.9451155 0.03198024         10       1023         0
#> 163     4        13       0.9674459 0.03198024          6        127         0
#> 164     4        14       0.8276137 0.03198024          4         31         0
#> 165     4        15       0.9805136 0.03198024          7        191         0
#> 166     4        16       0.9924236 0.03198024         10       1023         0
#> 167     4        17       0.8430639 0.03198024          8        511         0
#> 168     4        18       0.9988608 0.03198024          9        639         0
#> 169     4        19       0.9895100 0.03198024          4         15         0
#> 170     4        20       0.9769122 0.03198024          6        127         0
#> 171     4        21       0.9130575 0.03198024         10       1023         0
#> 172     4        22       0.9980508 0.03198024          8        383         0
#> 173     4        23       0.9990285 0.03198024          6         63         0
#> 174     4        24       0.9353453 0.03198024         10       1023         0
#> 175     4        25       0.9412026 0.03198024          7        127         0
#> 176     4        26       0.9982789 0.03198024          6         63         0
#> 177     4        27       0.9018996 0.03198024         10       1023         0
#> 178     4        28       0.9340106 0.03198024          9        703         0
#> 179     4        29       0.9981468 0.03198024          9        575         0
#> 180     4        30       0.9789302 0.03198024          9        831         0
#> 181     4        31       0.8868699 0.03198024          7        191         0
#> 182     4        32       0.9924879 0.03198024          9        895         0
#> 183     4        33       0.9482686 0.03198024          8        383         0
#> 184     4        34       0.9959221 0.03198024         10       1023         0
#> 185     4        35       0.9914571 0.03198024         10       1023         0
#> 186     4        36       0.9930731 0.03198024          6         63         0
#> 187     4        37       0.9605904 0.03198024          6         95         0
#> 188     4        38       0.9972483 0.03198024         10       1023         0
#> 189     4        39       0.9987651 0.03198024          9        575         0
#> 190     4        40       0.9921278 0.03198024          7        255         0
#> 191     4        41       0.9497307 0.03198024          6         63         0
#> 192     4        42       0.8972034 0.03198024         10       1023         0
#> 193     4        43       0.9761140 0.03198024          7        191         0
#> 194     4        44       0.8666288 0.03198024         10       1023         0
#> 195     4        45       0.9879294 0.03198024          7        255         0
#> 196     4        46       0.9948571 0.03198024          7        127         0
#> 197     4        47       0.9944272 0.03198024          6        127         0
#> 198     4        48       0.8983594 0.03198024          5         47         0
#> 199     4        49       0.9890020 0.03198024          9        703         0
#> 200     4        50       0.9850923 0.03198024          8        319         0
#>       Energy LogPosterior
#> 1   83.17058    -79.71523
#> 2   88.46335    -80.30620
#> 3   85.52624    -80.08154
#> 4   83.28521    -78.38110
#> 5   81.08484    -77.74140
#> 6   87.45117    -82.58233
#> 7   91.30524    -83.27058
#> 8   86.02754    -80.44703
#> 9   86.84384    -79.45870
#> 10  83.57325    -78.98902
#> 11  85.43327    -80.87561
#> 12  84.84325    -76.60463
#> 13  79.71992    -76.55103
#> 14  78.53293    -76.38145
#> 15  83.45874    -77.58010
#> 16  84.57260    -78.98299
#> 17  81.55407    -79.40978
#> 18  80.67639    -76.29762
#> 19  78.17347    -75.76636
#> 20  77.39442    -76.01737
#> 21  77.38390    -75.14359
#> 22  79.43180    -76.43350
#> 23  80.52547    -77.20222
#> 24  79.63132    -74.84074
#> 25  76.82266    -75.70538
#> 26  82.36136    -77.58170
#> 27  86.62572    -81.61899
#> 28  87.02138    -76.74058
#> 29  80.40361    -76.01881
#> 30  80.27587    -78.48493
#> 31  84.48328    -78.98499
#> 32  80.52157    -78.10883
#> 33  79.44776    -78.20614
#> 34  81.13325    -78.86615
#> 35  82.38078    -79.58949
#> 36  88.62434    -81.57258
#> 37  83.08007    -77.44769
#> 38  82.10544    -76.01642
#> 39  81.58161    -76.94775
#> 40  81.28914    -75.41048
#> 41  81.02919    -77.90187
#> 42  81.14215    -78.33017
#> 43  84.56517    -78.96803
#> 44  81.33338    -79.42271
#> 45  82.92329    -77.03712
#> 46  80.29354    -78.62697
#> 47  82.72882    -76.79732
#> 48  83.00383    -78.86617
#> 49  80.91825    -78.05679
#> 50  83.75800    -79.14299
#> 51  79.51911    -76.73780
#> 52  80.26962    -76.53840
#> 53  78.57539    -75.85698
#> 54  79.85391    -77.12497
#> 55  79.08280    -75.48862
#> 56  78.19671    -75.05014
#> 57  77.52590    -75.30450
#> 58  79.96528    -75.42032
#> 59  79.40290    -76.86514
#> 60  84.96301    -77.73463
#> 61  83.99106    -80.10425
#> 62  85.55683    -79.82521
#> 63  82.62949    -78.88119
#> 64  84.65243    -77.65041
#> 65  82.45828    -77.73788
#> 66  81.51115    -77.32974
#> 67  84.34726    -78.83837
#> 68  83.82577    -79.19763
#> 69  83.42174    -78.30920
#> 70  79.91034    -77.33826
#> 71  82.44350    -77.46893
#> 72  83.09450    -81.22025
#> 73  87.97554    -78.89796
#> 74  80.13757    -78.25308
#> 75  80.51761    -75.75648
#> 76  81.34145    -77.16128
#> 77  83.04645    -78.22539
#> 78  83.23757    -78.99184
#> 79  89.24184    -78.94422
#> 80  81.01373    -76.96020
#> 81  80.70807    -78.69514
#> 82  82.59751    -78.42229
#> 83  82.33668    -78.52027
#> 84  80.02112    -78.77968
#> 85  81.79308    -77.25647
#> 86  80.93072    -76.88755
#> 87  79.70899    -77.88894
#> 88  85.91497    -79.36578
#> 89  85.21527    -79.55737
#> 90  85.63426    -79.01196
#> 91  85.64356    -79.13415
#> 92  83.89845    -80.55743
#> 93  82.98120    -79.56202
#> 94  88.54700    -79.87825
#> 95  84.07368    -79.20598
#> 96  80.65840    -77.68990
#> 97  79.07463    -76.71678
#> 98  79.00585    -77.52144
#> 99  83.24576    -78.50062
#> 100 83.42481    -77.82666
#> 101 84.76804    -79.24448
#> 102 84.26949    -79.27212
#> 103 83.65149    -79.75327
#> 104 85.11140    -81.15469
#> 105 83.51631    -79.52246
#> 106 83.38415    -80.01177
#> 107 84.46703    -79.16333
#> 108 82.55377    -77.05955
#> 109 79.12061    -76.30476
#> 110 88.62101    -83.56752
#> 111 85.96441    -84.17524
#> 112 86.30705    -78.27332
#> 113 86.01486    -81.09507
#> 114 84.76599    -80.28450
#> 115 82.11015    -80.99296
#> 116 91.44807    -80.95033
#> 117 90.28163    -84.45787
#> 118 86.54228    -78.02431
#> 119 84.95764    -81.34716
#> 120 85.83230    -82.02398
#> 121 85.93277    -80.66813
#> 122 85.09975    -77.61853
#> 123 87.70308    -82.86321
#> 124 86.39953    -79.03526
#> 125 83.57058    -77.70008
#> 126 83.22348    -81.61913
#> 127 82.93997    -80.03762
#> 128 86.54244    -80.02183
#> 129 82.32140    -79.32482
#> 130 83.30393    -79.14053
#> 131 86.17931    -83.13445
#> 132 90.36419    -87.71828
#> 133 90.63199    -85.14706
#> 134 87.49552    -84.68754
#> 135 89.31409    -85.11212
#> 136 89.43842    -83.80957
#> 137 87.86388    -84.34496
#> 138 87.80721    -84.20389
#> 139 87.74069    -79.23641
#> 140 82.39571    -78.01253
#> 141 83.41796    -78.66656
#> 142 85.94575    -80.64826
#> 143 84.41087    -82.81183
#> 144 86.37056    -78.79323
#> 145 83.29139    -78.96700
#> 146 81.72538    -77.98287
#> 147 80.71585    -76.89289
#> 148 80.88727    -77.83812
#> 149 82.81873    -77.50018
#> 150 82.08382    -80.52659
#> 151 88.57406    -79.29047
#> 152 83.44910    -78.09141
#> 153 80.09167    -77.65424
#> 154 81.81552    -79.26496
#> 155 80.29263    -76.38105
#> 156 80.39227    -76.23212
#> 157 80.89092    -77.14047
#> 158 83.62910    -80.80563
#> 159 83.24108    -81.10400
#> 160 87.02213    -78.67041
#> 161 81.14315    -78.13905
#> 162 82.65370    -78.67945
#> 163 83.92355    -78.54259
#> 164 86.13173    -79.50173
#> 165 81.75264    -78.48940
#> 166 82.04444    -75.95749
#> 167 80.85561    -78.31486
#> 168 89.61268    -85.39934
#> 169 87.55132    -83.42507
#> 170 87.08758    -78.97180
#> 171 83.87443    -78.25577
#> 172 81.85453    -79.40039
#> 173 82.93052    -77.71615
#> 174 82.81442    -80.46143
#> 175 83.81949    -78.83272
#> 176 80.58618    -78.78535
#> 177 89.45394    -75.96623
#> 178 81.72626    -78.28802
#> 179 79.99585    -75.75171
#> 180 83.74128    -77.67358
#> 181 82.98996    -78.39464
#> 182 81.04839    -76.79816
#> 183 79.58438    -76.46149
#> 184 80.76106    -76.22015
#> 185 81.03834    -78.97766
#> 186 80.94508    -77.64284
#> 187 79.19824    -78.19956
#> 188 82.13783    -78.78973
#> 189 81.47872    -76.65154
#> 190 81.42264    -78.07513
#> 191 82.51235    -80.38530
#> 192 87.95338    -82.35993
#> 193 83.81848    -81.26266
#> 194 87.95459    -83.29291
#> 195 85.23120    -78.88481
#> 196 81.49128    -77.30369
#> 197 81.32215    -77.09604
#> 198 82.69018    -78.23515
#> 199 83.57473    -80.06101
#> 200 81.77902    -79.14845

# }