distribution {bayestestR}	R Documentation

Empirical Distributions

Description

Generate a sequence of n-quantiles, i.e., a sample of size n with a near-perfect distribution.

Usage

distribution(type = "normal", ...)

distribution_normal(n, mean = 0, sd = 1, random = FALSE, ...)

distribution_binomial(n, size = 1, prob = 0.5, random = FALSE, ...)

distribution_cauchy(n, location = 0, scale = 1, random = FALSE, ...)

distribution_poisson(n, lambda = 1, random = FALSE, ...)

distribution_student(n, df, ncp, random = FALSE, ...)

distribution_chisquared(n, df, ncp = 0, random = FALSE, ...)

distribution_uniform(n, min = 0, max = 1, random = FALSE, ...)

distribution_beta(n, shape1, shape2, ncp = 0, random = FALSE, ...)

distribution_tweedie(n, xi = NULL, mu, phi, power = NULL, random = FALSE, ...)

distribution_gamma(n, shape, scale = 1, random = FALSE, ...)

distribution_custom(n, type = "norm", ..., random = FALSE)

distribution_mixture_normal(n, mean = c(-3, 3), sd = 1, random = FALSE, ...)

rnorm_perfect(n, mean = 0, sd = 1)

Arguments

type
Can be any of the names from base R's Distributions, like "cauchy", "pois" or "beta".

...
Arguments passed to or from other methods.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

mean
vector of means.

sd
vector of standard deviations.

random
Generate near-perfect or random (simple wrappers for the base R r* functions) distributions.

size
number of trials (zero or more).

prob
probability of success on each trial.

location
location and scale parameters.

scale
location and scale parameters.

lambda
vector of (non-negative) means.

df
degrees of freedom (> 0, maybe non-integer). df = Inf is allowed.

ncp
non-centrality parameter delta; currently except for rt(), only for abs(ncp) <= 37.62. If omitted, use the central t distribution.

min
lower and upper limits of the distribution. Must be finite.

max
lower and upper limits of the distribution. Must be finite.

shape1
non-negative parameters of the Beta distribution.

shape2
non-negative parameters of the Beta distribution.

xi
the value of xi such that the variance is var(Y) = phi * mu^xi

mu
the mean

phi
the dispersion

power
a synonym for xi

shape
shape and scale parameters. Must be positive, scale strictly.

Examples

library(bayestestR)
x <- distribution(n = 10)
plot(density(x))

x <- distribution(type = "gamma", n = 100, shape = 2)
plot(density(x))

[Package bayestestR version 0.7.0 Index]