Test for Practical Equivalence

Perform a Test for Practical Equivalence for Bayesian and frequentist models.

equivalence_test(x, ...)

# S3 method for default
equivalence_test(x, ...)

# S3 method for numeric
equivalence_test(x, range = "default", ci = 0.89,
  verbose = TRUE, ...)

# S3 method for data.frame
equivalence_test(x, range = "default", ci = 0.89,
  verbose = TRUE, ...)

# S3 method for emmGrid
equivalence_test(x, range = "default", ci = 0.89,
  verbose = TRUE, ...)

# S3 method for BFBayesFactor
equivalence_test(x, range = "default",
  ci = 0.89, verbose = TRUE, ...)

# S3 method for stanreg
equivalence_test(x, range = "default", ci = 0.89,
  parameters = NULL, verbose = TRUE, ...)

# S3 method for brmsfit
equivalence_test(x, range = "default", ci = 0.89,
  parameters = NULL, verbose = TRUE, ...)

Arguments

x
Vector representing a posterior distribution. Can also be a stanreg or brmsfit model.

...
Currently not used.

range
ROPE's lower and higher bounds. Should be a vector of length two (e.g., c(-0.1, 0.1)) or "default". If "default", the range is set to c(-0.1, 0.1) if input is a vector, and based on rope_range() if a Bayesian model is provided.

ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

verbose
Toggle off warnings.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

Value

A data frame with following columns:

Parameter The model parameter(s), if x is a model-object. If x is a vector, this column is missing.

CI The probability of the HDI.

ROPE_low, ROPE_high The limits of the ROPE. These values are identical for all parameters.

ROPE_Percentage The proportion of the HDI that lies inside the ROPE.

ROPE_Equivalence The "test result", as character. Either "rejected", "accepted" or "undecided".

HDI_low , HDI_high The lower and upper HDI limits for the parameters.

Details

Documentation is accessible for:

Bayesian models

Frequentist models

For Bayesian models, the Test for Practical Equivalence is based on the "HDI+ROPE decision rule" (Kruschke, 2014, 2018) to check whether parameter values should be accepted or rejected against an explicitly formulated "null hypothesis" (i.e., a ROPE). In other words, it checks the percentage of the 89% HDI that is the null region (the ROPE). If this percentage is sufficiently low, the null hypothesis is rejected. If this percentage is sufficiently high, the null hypothesis is accepted.

Using the ROPE and the HDI, Kruschke (2018) suggests using the percentage of the 95% (or 89%, considered more stable) HDI that falls within the ROPE as a decision rule. If the HDI is completely outside the ROPE, the "null hypothesis" for this parameter is "rejected". If the ROPE completely covers the HDI, i.e., all most credible values of a parameter are inside the region of practical equivalence, the null hypothesis is accepted. Else, it’s undecided whether to accept or reject the null hypothesis. If the full ROPE is used (i.e., 100% of the HDI), then the null hypothesis is rejected or accepted if the percentage of the posterior within the ROPE is smaller than to 2.5% or greater than 97.5%. Desirable results are low proportions inside the ROPE (the closer to zero the better).

Some attention is required for finding suitable values for the ROPE limits (argument range). See 'Details' in rope_range() for further information.

Multicollinearity: Non-independent covariates

When parameters show strong correlations, i.e. when covariates are not independent, the joint parameter distributions may shift towards or away from the ROPE. In such cases, the test for practical equivalence may have inappropriate results. Collinearity invalidates ROPE and hypothesis testing based on univariate marginals, as the probabilities are conditional on independence. Most problematic are the results of the "undecided" parameters, which may either move further towards "rejection" or away from it (Kruschke 2014, 340f).

equivalence_test() performs a simple check for pairwise correlations between parameters, but as there can be collinearity between more than two variables, a first step to check the assumptions of this hypothesis testing is to look at different pair plots. An even more sophisticated check is the projection predictive variable selection (Piironen and Vehtari 2017).

Note

There is a print()-method with a digits-argument to control the amount of digits in the output, and there is a plot()-method to visualize the results from the equivalence-test (for models only).

References

Kruschke, J. K. (2018). Rejecting or accepting parameter values in Bayesian estimation. Advances in Methods and Practices in Psychological Science, 1(2), 270-280. doi: 10.1177/2515245918771304

Kruschke, J. (2014). Doing Bayesian data analysis: A tutorial with R, JAGS, and Stan. Academic Press

Piironen, J., & Vehtari, A. (2017). Comparison of Bayesian predictive methods for model selection. Statistics and Computing, 27(3), 711–735. doi: 10.1007/s11222-016-9649-y

Examples

library(bayestestR)

equivalence_test(x = rnorm(1000, 0, 0.01), range = c(-0.1, 0.1))

#> # Test for Practical Equivalence
#> 
#>   ROPE: [-0.10 0.10]
#> 
#>        H0 inside ROPE      89% HDI
#>  Accepted      1.00 % [-0.01 0.02]
#> 

equivalence_test(x = rnorm(1000, 0, 1), range = c(-0.1, 0.1))

#> # Test for Practical Equivalence
#> 
#>   ROPE: [-0.10 0.10]
#> 
#>         H0 inside ROPE      89% HDI
#>  Undecided      0.09 % [-1.67 1.49]
#> 

equivalence_test(x = rnorm(1000, 1, 0.01), range = c(-0.1, 0.1))

#> # Test for Practical Equivalence
#> 
#>   ROPE: [-0.10 0.10]
#> 
#>        H0 inside ROPE     89% HDI
#>  Rejected      0.00 % [0.99 1.02]
#> 

equivalence_test(x = rnorm(1000, 1, 1), ci = c(.50, .99))

#> # Test for Practical Equivalence
#> 
#>   ROPE: [-0.10 0.10]
#> 
#>         H0 inside ROPE      50% HDI
#>  Undecided      0.05 % [ 0.04 1.32]
#> 
#>         H0 inside ROPE      99% HDI
#>  Undecided      0.06 % [-1.73 3.64]
#> 

# print more digits
test <- equivalence_test(x = rnorm(1000, 1, 1), ci = c(.50, .99))
print(test, digits = 4)

#> # Test for Practical Equivalence
#> 
#>   ROPE: [-0.1000 0.1000]
#> 
#>        H0 inside ROPE          50% HDI
#>  Rejected    0.0000 % [ 0.4559 1.7549]
#> 
#>         H0 inside ROPE          99% HDI
#>  Undecided    0.0585 % [-1.5215 3.6113]
#> 

if (FALSE) {
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
equivalence_test(model)
equivalence_test(model, ci = c(.50, 1))

# plot result
test <- equivalence_test(model)
plot(test)

library(emmeans)
equivalence_test(emtrends(model, ~1, "wt"))

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
equivalence_test(model)
equivalence_test(model, ci = c(.50, .99))

ibrary(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
equivalence_test(bf)
equivalence_test(bf, ci = c(.50, .99))
}