Test for Practical Equivalence

Perform a Test for Practical Equivalence based on the "HDI+ROPE decision rule" (Kruschke, 2018) to check whether parameter values should be accepted or rejected against an explicitly formulated "null hypothesis".

equivalence_test(posterior, bounds = "default", ci = 0.9,
  verbose = TRUE)

Arguments

posterior
vector representing a posterior distribution. Can also be a stanreg or brmsfit model.

bounds
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to c(0.1, 0.1) if input is a vector and x +- 0.1*SD(response) if a Bayesian model is provided.

ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use.

verbose
Toggle off warnings.

Details

Using the ROPE and the HDI, Kruschke (2010, 2011, 2014, 2018) suggest using the percentage of the 95% (or 90%, considered more stable) HDI that falls within the ROPE as a decision rule. If the HDI is completely outside the ROPE, the "null hypothesis" for this parameter is "rejected". If the ROPE completely covers the HDI, i.e. all most credible values of a parameter are inside the region of practical equivalence, the null hypothesis is accepted. Else, it’s undecided whether to accept or reject the null hypothesis.

References

sjstats

Examples

library(bayestestR)

equivalence_test(posterior = rnorm(1000, 0, 0.01), bounds = c(-0.1, 0.1))

#>   CI ROPE_low ROPE_high ROPE_Percentage ROPE_Equivalence
#> 1 90     -0.1       0.1             100         accepted

equivalence_test(posterior = rnorm(1000, 0, 1), bounds = c(-0.1, 0.1))

#>   CI ROPE_low ROPE_high ROPE_Percentage ROPE_Equivalence
#> 1 90     -0.1       0.1        10.54384        undecided

equivalence_test(posterior = rnorm(1000, 1, 0.01), bounds = c(-0.1, 0.1))

#>   CI ROPE_low ROPE_high ROPE_Percentage ROPE_Equivalence
#> 1 90     -0.1       0.1               0         rejected

equivalence_test(posterior = rnorm(1000, 1, 1), ci = c(.50, .99))

#>   CI ROPE_low ROPE_high ROPE_Percentage ROPE_Equivalence
#> 1 50     -0.1       0.1        0.000000         rejected
#> 2 99     -0.1       0.1        5.449041        undecided
# NOT RUN {
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
equivalence_test(model)
equivalence_test(model, ci = c(.50, 1))

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
equivalence_test(model)
equivalence_test(model, ci = c(.50, .99))
# }