Highest Density Interval (HDI)

Compute the Highest Density Interval (HDI) of a posterior distribution, i.e., all points within the interval have a higher probability density than points outside the interval. The HDI can be used in the context of Bayesian posterior characterisation as Credible Interval (CI).

hdi(x, ...)

# S3 method for numeric
hdi(x, ci = 0.9, verbose = TRUE, ...)

# S3 method for stanreg
hdi(x, ci = 0.9, effects = c("fixed", "random",
  "all"), parameters = NULL, verbose = TRUE, ...)

# S3 method for brmsfit
hdi(x, ci = 0.9, effects = c("fixed", "random",
  "all"), component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL, verbose = TRUE, ...)

Arguments

x
Vector representing a posterior distribution. Can also be a stanreg or brmsfit model.

...
Currently not used.

ci
Value or vector of probability of the interval (between 0 and 1) to be estimated. Named Credible Interval (CI) for consistency.

verbose
Toggle off warnings.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Value

A data frame with following columns:

Parameter The model parameter(s), if x is a model-object. If x is a vector, this column is missing.

CI The probability of the HDI.

CI_low , CI_high The lower and upper HDI limits for the parameters.

Details

Unlike equal-tailed intervals (see ci) that typically exclude 2.5% from each tail of the distribution, the HDI is not equal-tailed and therefore always includes the mode(s) of posterior distributions.

By default, hdi() returns the 90% intervals (ci = 0.9), deemed to be more stable than, for instance, 95% intervals (Kruschke, 2015). An effective sample size of at least 10.000 is recommended if 95% intervals should be computed (Kruschke, 2015, p. 183ff).

References

Kruschke, J. (2015). Doing Bayesian data analysis: A tutorial with R, JAGS, and Stan. Academic Press.

Examples

library(bayestestR)

posterior <- rnorm(1000)
hdi(posterior, ci = .90)

#> # Highest Density Interval
#> 
#>       90% HDI
#>  [-1.48 1.89]
#> 

hdi(posterior, ci = c(.80, .90, .95))

#> # Highest Density Intervals
#> 
#>       80% HDI
#>  [-1.24 1.31]
#> 
#> 
#>       90% HDI
#>  [-1.48 1.89]
#> 
#> 
#>       95% HDI
#>  [-1.83 2.08]
#> 
#> 
# NOT RUN {
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
hdi(model)
hdi(model, ci = c(.80, .90, .95))

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
hdi(model)
hdi(model, ci = c(.80, .90, .95))
# }