Highest Density Interval (HDI)

Compute the Highest Density Interval (HDI) of posterior distributions. All points within this interval have a higher probability density than points outside the interval. The HDI can be used in the context of uncertainty characterisation of posterior distributions as Credible Interval (CI).

hdi(x, ...)

# S3 method for numeric
hdi(x, ci = 0.89, verbose = TRUE, ...)

# S3 method for data.frame
hdi(x, ci = 0.89, verbose = TRUE, ...)

# S3 method for emmGrid
hdi(x, ci = 0.89, verbose = TRUE, ...)

# S3 method for stanreg
hdi(x, ci = 0.89, effects = c("fixed", "random",
  "all"), parameters = NULL, verbose = TRUE, ...)

# S3 method for brmsfit
hdi(x, ci = 0.89, effects = c("fixed", "random",
  "all"), component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL, verbose = TRUE, ...)

# S3 method for BFBayesFactor
hdi(x, ci = 0.89, verbose = TRUE, ...)

Arguments

x
Vector representing a posterior distribution. Can also be a stanreg, brmsfit or a BayesFactor model.

...
Currently not used.

ci
Value or vector of probability of the (credible) interval - CI (between 0 and 1) to be estimated. Default to .89 (89%).

verbose
Toggle off warnings.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Value

A data frame with following columns:

Parameter The model parameter(s), if x is a model-object. If x is a vector, this column is missing.

CI The probability of the credible interval.

CI_low, CI_high The lower and upper credible interval limits for the parameters.

Details

Unlike equal-tailed intervals (see eti()) that typically exclude 2.5% from each tail of the distribution and always include the median, the HDI is not equal-tailed and therefore always includes the mode(s) of posterior distributions.

By default, hdi() and eti() returns the 89% intervals (ci = 0.89), deemed to be more stable than, for instance, 95% intervals (Kruschke, 2014). An effective sample size of at least 10.000 is recommended if 95% intervals should be computed (Kruschke, 2014, p. 183ff). Moreover, 89 is the not exceed the already unstable and arbitrary 95% threshold (McElreath, 2015).

A 90% equal-tailed interval (ETI) has 5% of the distribution on either side of its limits. It indicates the 5th percentile and the 95h percentile. In symmetric distributions, the two methods of computing credible intervals, the ETI and the HDI, return similar results.

This is not the case for skewed distributions. Indeed, it is possible that parameter values in the ETI have lower credibility (are less probable) than parameter values outside the ETI. This property seems undesirable as a summary of the credible values in a distribution.

On the other hand, the ETI range does change when transformations are applied to the distribution (for instance, for a log odds scale to probabilities): the lower and higher bounds of the transformed distribution will correspond to the transformed lower and higher bounds of the original distribution. On the contrary, applying transformations to the distribution will change the resulting HDI.

References

Kruschke, J. (2014). Doing Bayesian data analysis: A tutorial with R, JAGS, and Stan. Academic Press.

McElreath, R. (2015). Statistical rethinking: A Bayesian course with examples in R and Stan. Chapman and Hall/CRC.

Examples

library(bayestestR)

posterior <- rnorm(1000)
hdi(posterior, ci = .89)

#> # Highest Density Interval
#> 
#>        89% HDI
#>  [-1.63, 1.60]
#> 

hdi(posterior, ci = c(.80, .90, .95))

#> # Highest Density Intervals
#> 
#>        80% HDI
#>  [-1.25, 1.27]
#> 
#> 
#>        90% HDI
#>  [-1.61, 1.71]
#> 
#> 
#>        95% HDI
#>  [-2.17, 1.86]
#> 
#> 

df <- data.frame(replicate(4, rnorm(100)))
hdi(df)

#>   Parameter CI    CI_low  CI_high
#> 1        X1 89 -1.661651 1.561213
#> 2        X2 89 -1.406116 1.826320
#> 3        X3 89 -1.533729 1.663738
#> 4        X4 89 -1.626538 1.506512

hdi(df, ci = c(.80, .90, .95))

#>    Parameter CI    CI_low  CI_high
#> 1         X1 80 -1.448216 1.404471
#> 2         X1 90 -1.630196 1.703493
#> 3         X1 95 -2.026692 1.703493
#> 4         X2 80 -1.406116 1.217964
#> 5         X2 90 -1.653797 1.653508
#> 6         X2 95 -1.653797 2.150041
#> 7         X3 80 -1.637334 1.001949
#> 8         X3 90 -1.533729 1.681119
#> 9         X3 95 -1.882425 2.016106
#> 10        X4 80 -1.521842 1.030051
#> 11        X4 90 -1.626538 1.633639
#> 12        X4 95 -1.626538 1.969543

library(rstanarm)
model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 15
#> Chain 1:            adapt_window = 75
#> Chain 1:            term_buffer = 10
#> Chain 1: 
#> Chain 1: Iteration:   1 / 200 [  0%]  (Warmup)
#> Chain 1: Iteration:  20 / 200 [ 10%]  (Warmup)
#> Chain 1: Iteration:  40 / 200 [ 20%]  (Warmup)
#> Chain 1: Iteration:  60 / 200 [ 30%]  (Warmup)
#> Chain 1: Iteration:  80 / 200 [ 40%]  (Warmup)
#> Chain 1: Iteration: 100 / 200 [ 50%]  (Warmup)
#> Chain 1: Iteration: 101 / 200 [ 50%]  (Sampling)
#> Chain 1: Iteration: 120 / 200 [ 60%]  (Sampling)
#> Chain 1: Iteration: 140 / 200 [ 70%]  (Sampling)
#> Chain 1: Iteration: 160 / 200 [ 80%]  (Sampling)
#> Chain 1: Iteration: 180 / 200 [ 90%]  (Sampling)
#> Chain 1: Iteration: 200 / 200 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.012 seconds (Warm-up)
#> Chain 1:                0.013 seconds (Sampling)
#> Chain 1:                0.025 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: WARNING: There aren't enough warmup iterations to fit the
#> Chain 2:          three stages of adaptation as currently configured.
#> Chain 2:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 2:          the given number of warmup iterations:
#> Chain 2:            init_buffer = 15
#> Chain 2:            adapt_window = 75
#> Chain 2:            term_buffer = 10
#> Chain 2: 
#> Chain 2: Iteration:   1 / 200 [  0%]  (Warmup)
#> Chain 2: Iteration:  20 / 200 [ 10%]  (Warmup)
#> Chain 2: Iteration:  40 / 200 [ 20%]  (Warmup)
#> Chain 2: Iteration:  60 / 200 [ 30%]  (Warmup)
#> Chain 2: Iteration:  80 / 200 [ 40%]  (Warmup)
#> Chain 2: Iteration: 100 / 200 [ 50%]  (Warmup)
#> Chain 2: Iteration: 101 / 200 [ 50%]  (Sampling)
#> Chain 2: Iteration: 120 / 200 [ 60%]  (Sampling)
#> Chain 2: Iteration: 140 / 200 [ 70%]  (Sampling)
#> Chain 2: Iteration: 160 / 200 [ 80%]  (Sampling)
#> Chain 2: Iteration: 180 / 200 [ 90%]  (Sampling)
#> Chain 2: Iteration: 200 / 200 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.016 seconds (Warm-up)
#> Chain 2:                0.014 seconds (Sampling)
#> Chain 2:                0.03 seconds (Total)
#> Chain 2: 

#> Warning: The largest R-hat is 1.06, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

hdi(model)

#> # Highest Density Interval
#> 
#>    Parameter        89% HDI
#>  (Intercept) [29.79, 45.19]
#>           wt [-6.54, -4.28]
#>         gear [-1.55,  1.35]
#> 

hdi(model, ci = c(.80, .90, .95))

#> # Highest Density Intervals
#> 
#>    Parameter        80% HDI
#>  (Intercept) [32.15, 44.86]
#>           wt [-6.43, -4.73]
#>         gear [-1.20,  1.16]
#> 
#> 
#>    Parameter        90% HDI
#>  (Intercept) [30.62, 46.14]
#>           wt [-6.51, -4.17]
#>         gear [-1.55,  1.50]
#> 
#> 
#>    Parameter        95% HDI
#>  (Intercept) [28.55, 48.16]
#>           wt [-6.82, -4.03]
#>         gear [-2.07,  1.50]
#> 
#> 

library(emmeans)
hdi(emtrends(model, ~1, "wt"))

#>   Parameter CI   CI_low   CI_high
#> 1   overall 89 -6.54033 -4.279254
# NOT RUN {
library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
hdi(model)
hdi(model, ci = c(.80, .90, .95))

library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
hdi(bf)
hdi(bf, ci = c(.80, .90, .95))
# }