Probability of not being in ROPE

Compute the proportion of the posterior distribution that doesn't lie within a region of practical equivalence (ROPE). It is equivalent to running rope(..., ci = 1).

p_rope(x, ...)

# S3 method for default
p_rope(x, ...)

# S3 method for numeric
p_rope(x, range = "default", ...)

# S3 method for data.frame
p_rope(x, range = "default", ...)

# S3 method for emmGrid
p_rope(x, range = "default", ...)

# S3 method for BFBayesFactor
p_rope(x, range = "default", ...)

# S3 method for MCMCglmm
p_rope(x, range = "default", ...)

# S3 method for stanreg
p_rope(
  x,
  range = "default",
  effects = c("fixed", "random", "all"),
  parameters = NULL,
  ...
)

# S3 method for brmsfit
p_rope(
  x,
  range = "default",
  effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL,
  ...
)

Arguments

x
Vector representing a posterior distribution. Can also be a stanreg or brmsfit model.

...
Currently not used.

range
ROPE's lower and higher bounds. Should be a vector of length two (e.g., c(-0.1, 0.1)) or "default". If "default", the range is set to c(-0.1, 0.1) if input is a vector, and based on rope_range() if a Bayesian model is provided.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Examples

library(bayestestR)

p_rope(x = rnorm(1000, 0, 0.01), range = c(-0.1, 0.1))

#>   ROPE_low ROPE_high p_ROPE
#> 1     -0.1       0.1      1

p_rope(x = mtcars, range = c(-0.1, 0.1))

#>    Parameter ROPE_low ROPE_high p_ROPE
#> 1        mpg     -0.1       0.1   0.00
#> 2        cyl     -0.1       0.1   0.00
#> 3       disp     -0.1       0.1   0.00
#> 4         hp     -0.1       0.1   0.00
#> 5       drat     -0.1       0.1   0.00
#> 6         wt     -0.1       0.1   0.00
#> 7       qsec     -0.1       0.1   0.00
#> 8         vs     -0.1       0.1   0.56
#> 9         am     -0.1       0.1   0.59
#> 10      gear     -0.1       0.1   0.00
#> 11      carb     -0.1       0.1   0.00