Practical Significance (ps)

Compute the probability of Practical Significance (ps), which can be conceptualized as a unidirectional equivalence test. It returns the probability that effect is above a given threshold corresponding to a negligible effect in the median's direction. Mathematically, it is defined as the proportion of the posterior distribution of the median sign above the threshold.

p_significance(x, ...)

# S3 method for numeric
p_significance(x, threshold = "default", ...)

# S3 method for emmGrid
p_significance(x, threshold = "default", ...)

# S3 method for stanreg
p_significance(
  x,
  threshold = "default",
  effects = c("fixed", "random", "all"),
  parameters = NULL,
  verbose = TRUE,
  ...
)

# S3 method for brmsfit
p_significance(
  x,
  threshold = "default",
  effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL,
  verbose = TRUE,
  ...
)

Arguments

x
Vector representing a posterior distribution. Can also be a stanreg or brmsfit model.

...
Currently not used.

threshold
The threshold value that separates significant from negligible effect. If "default", the range is set to 0.1 if input is a vector, and based on rope_range() if a Bayesian model is provided.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

verbose
Toggle off warnings.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

Value

Values between 0 and 1 corresponding to the probability of practical significance (ps).

Details

p_significance() returns the proportion of a probability distribution (x) that is outside a certain range (the negligible effect, or ROPE, see argument threshold). If there are values of the distribution both below and above the ROPE, p_significance() returns the higher probability of a value being outside the ROPE. Typically, this value should be larger than 0.5 to indicate practical significance. However, if the range of the negligible effect is rather large compared to the range of the probability distribution x, p_significance() will be less than 0.5, which indicates no clear practical significance.

Examples

library(bayestestR)

# Simulate a posterior distribution of mean 1 and SD 1
# ----------------------------------------------------
posterior <- rnorm(1000, mean = 1, sd = 1)
p_significance(posterior)

#> ps [0.10] = 80.10%

# Simulate a dataframe of posterior distributions
# -----------------------------------------------
df <- data.frame(replicate(4, rnorm(100)))
p_significance(df)

#> # Probability of Significance (ps [0.10])
#> 
#> Parameter |     ps
#> ------------------
#> X1        | 51.00%
#> X2        | 55.00%
#> X3        | 47.00%
#> X4        | 47.00%
#> 

if (FALSE) {
# rstanarm models
# -----------------------------------------------
if (require("rstanarm")) {
  model <- rstanarm::stan_glm(mpg ~ wt + cyl,
    data = mtcars,
    chains = 2, refresh = 0
  )
  p_significance(model)
}
}