Point-estimates of posterior distributions

Compute various point-estimates, such as the mean, the median or the MAP, to describe posterior distributions.

point_estimate(x, centrality = "median", dispersion = FALSE, ...)

# S3 method for stanreg
point_estimate(x, centrality = "median",
  dispersion = FALSE, effects = c("fixed", "random", "all"),
  parameters = NULL, ...)

# S3 method for brmsfit
point_estimate(x, centrality = "median",
  dispersion = FALSE, effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL, ...)

# S3 method for BFBayesFactor
point_estimate(x, centrality = "median",
  dispersion = FALSE, ...)

Arguments

x
Vector representing a posterior distribution. Can also be a stanreg or brmsfit model.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

...
Additional arguments to be passed to or from methods.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

References

Vignette In-Depth 1: Comparison of Point-Estimates

Examples

library(bayestestR)

point_estimate(rnorm(1000))

#>       Median
#> 1 0.06233878

point_estimate(rnorm(1000), centrality = "all", dispersion = TRUE)

#>       Median      MAD        Mean        SD       MAP
#> 1 0.03110178 1.022674 0.005866268 0.9821226 0.3398239

point_estimate(rnorm(1000), centrality = c("median", "MAP"))

#>       Median         MAP
#> 1 0.03592228 -0.04653519

df <- data.frame(replicate(4, rnorm(100)))
point_estimate(df, centrality = "all", dispersion = TRUE)

#>   Parameter       Median       MAD        Mean        SD         MAP
#> 1        X1 -0.077717201 1.0140930 -0.03759236 0.9749721 -0.47703814
#> 2        X2 -0.045231103 0.9805633 -0.08963683 0.9212472 -0.03956969
#> 3        X3 -0.002583230 0.9564785 -0.03667868 0.9653942  0.25162819
#> 4        X4  0.002098652 0.9340898 -0.04519869 0.9158872  0.11879885

point_estimate(df, centrality = c("median", "MAP"))

#>   Parameter       Median         MAP
#> 1        X1 -0.077717201 -0.47703814
#> 2        X2 -0.045231103 -0.03956969
#> 3        X3 -0.002583230  0.25162819
#> 4        X4  0.002098652  0.11879885
# NOT RUN {
# rstanarm models
# -----------------------------------------------
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
point_estimate(model, centrality = "all", dispersion = TRUE)
point_estimate(model, centrality = c("median", "MAP"))

# emmeans estimates
# -----------------------------------------------
library(emmeans)
point_estimate(emtrends(model, ~1, "wt"), centrality = c("median", "MAP"))

# brms models
# -----------------------------------------------
library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
point_estimate(model, centrality = "all", dispersion = TRUE)
point_estimate(model, centrality = c("median", "MAP"))

# BayesFactor objects
# -----------------------------------------------
library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
point_estimate(bf, centrality = "all", dispersion = TRUE)
point_estimate(bf, centrality = c("median", "MAP"))
# }