Region of Practical Equivalence (ROPE)

Compute the proportion (in percentage) of the HDI (default to the 90% HDI) of a posterior distribution that lies within a region of practical equivalence.

rope(posterior, bounds = c(-0.1, 0.1), CI = 90, verbose = TRUE)

Arguments

posterior
vector representing a posterior distribution.

bounds
ROPE's lower and higher bounds.

CI
the credible interval to use.

verbose
toggle off warnings.

Details

Statistically, the probability of a posterior distribution of being different from 0 does not make much sense (the probability of it being different from a single point being infinite). Therefore, the idea underlining ROPE is to let the user define an area around the null value enclosing values that are equivalent to the null value for practical purposes (2010, 2011, 2014). Kruschke (2018) suggests that such null value could be set, by default, to the -0.1 to 0.1 range of a standardized parameter (negligible effect size according to Cohen). This could be generalized: For instance, for linear models, the ROPE could be set as 0 +/- .1 * sd(y). Kruschke (2010, 2011, 2014) suggest using the proportion of the 95% (or 90%, considered more stable) HDI that falls within the ROPE as an index for "null-hypothesis" testing (as understood under the Bayesian framework, see equivalence_test). Besides the ROPE-based decision criteria, the proportion of the 95% CI that falls in the ROPE can be used as a continuous index.

Examples

library(bayestestR)

rope(posterior = rnorm(1000, 0, 0.01), bounds = c(-0.1, 0.1))

#> 100.00% in ROPE

rope(posterior = rnorm(1000, 0, 1), bounds = c(-0.1, 0.1))

#> 8.10% in ROPE

rope(posterior = rnorm(1000, 1, 0.01), bounds = c(-0.1, 0.1))

#> 0.00% in ROPE

rope(posterior = rnorm(1000, 1, 1), CI = c(90, 95))

#> $CI_90
#> 5.77% in ROPE
#> $CI_95
#> 5.47% in ROPE