Region of Practical Equivalence (ROPE)

Compute the proportion (in percentage) of the HDI (default to the 90% HDI) of a posterior distribution that lies within a region of practical equivalence.

rope(x, ...)

# S3 method for numeric
rope(x, range = "default", ci = 0.9,
  verbose = TRUE, ...)

# S3 method for stanreg
rope(x, range = "default", ci = 0.9,
  effects = c("fixed", "random", "all"), parameters = NULL,
  verbose = TRUE, ...)

# S3 method for brmsfit
rope(x, range = "default", ci = 0.9,
  effects = c("fixed", "random", "all"), component = c("conditional",
  "zi", "zero_inflated", "all"), parameters = NULL, verbose = TRUE,
  ...)

Arguments

x
Vector representing a posterior distribution. Can also be a stanreg or brmsfit model.

...
Currently not used.

range
ROPE's lower and higher bounds. Should be a vector of length two (e.g., c(-0.1, 0.1)) or "default". If "default", the range is set to c(0.1, 0.1) if input is a vector, and x +- 0.1*SD(response) if a Bayesian model is provided.

ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use.

verbose
Toggle off warnings.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the modelbe returned? May be abbreviated. Only applies to brms-models.

Details

Statistically, the probability of a posterior distribution of being different from 0 does not make much sense (the probability of it being different from a single point being infinite). Therefore, the idea underlining ROPE is to let the user define an area around the null value enclosing values that are equivalent to the null value for practical purposes (Kruschke 2010, 2011, 2014).

Kruschke (2018) suggests that such null value could be set, by default, to the -0.1 to 0.1 range of a standardized parameter (negligible effect size according to Cohen, 1988). This could be generalized: For instance, for linear models, the ROPE could be set as 0 +/- .1 * sd(y). This ROPE range can be automatically computed for models using the rope_range function.

Kruschke (2010, 2011, 2014) suggests using the proportion of the 95% (or 90%, considered more stable) HDI that falls within the ROPE as an index for "null-hypothesis" testing (as understood under the Bayesian framework, see equivalence_test).

Examples

library(bayestestR)

rope(x = rnorm(1000, 0, 0.01), range = c(-0.1, 0.1))

#> # Proportion of samples inside the ROPE [-0.10, 0.10]:
#> 
#>  inside ROPE
#>     100.00 %
#> 

rope(x = rnorm(1000, 0, 1), range = c(-0.1, 0.1))

#> # Proportion of samples inside the ROPE [-0.10, 0.10]:
#> 
#>  inside ROPE
#>       9.21 %
#> 

rope(x = rnorm(1000, 1, 0.01), range = c(-0.1, 0.1))

#> # Proportion of samples inside the ROPE [-0.10, 0.10]:
#> 
#>  inside ROPE
#>       0.00 %
#> 

rope(x = rnorm(1000, 1, 1), ci = c(.90, .95))

#> # Proportions of samples inside the ROPE [-0.10, 0.10]:
#> 
#> ROPE for the 90% HDI:
#> 
#>  inside ROPE
#>       5.77 %
#> 
#> 
#> ROPE for the 95% HDI:
#> 
#>  inside ROPE
#>       5.47 %
#> 
#> 
# NOT RUN {
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
rope(model)
rope(model, ci = c(.90, .95))

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
rope(model)
rope(model, ci = c(.90, .95))
# }