Compute Support Intervals

A support interval contains only the values of the parameter that predict the observed data better than average, by some degree k; these are values of the parameter that are associated with an updating factor greater or equal than k. From the perspective of the Savage-Dickey Bayes factor, testing against a point null hypothesis for any value within the support interval will yield a Bayes factor smaller than 1/k.

For more info, in particular on specifying correct priors for factors with more than 2 levels, see the Bayes factors vignette.

si(posterior, prior = NULL, BF = 1, verbose = TRUE, ...)

# S3 method for numeric
si(posterior, prior = NULL, BF = 1, verbose = TRUE, ...)

# S3 method for stanreg
si(
  posterior,
  prior = NULL,
  BF = 1,
  verbose = TRUE,
  effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL,
  ...
)

# S3 method for brmsfit
si(
  posterior,
  prior = NULL,
  BF = 1,
  verbose = TRUE,
  effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL,
  ...
)

# S3 method for emmGrid
si(posterior, prior = NULL, BF = 1, verbose = TRUE, ...)

# S3 method for data.frame
si(posterior, prior = NULL, BF = 1, verbose = TRUE, ...)

Arguments

posterior
A numerical vector, stanreg / brmsfit object, emmGrid or a data frame - representing a posterior distribution(s) from (see 'Details').

prior
An object representing a prior distribution (see 'Details').

BF
The amount of support required to be included in the support interval.

verbose
Toggle off warnings.

...
Arguments passed to and from other methods.

effects
Should results for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

component
Should results for all parameters, parameters for the conditional model or the zero-inflated part of the model be returned? May be abbreviated. Only applies to brms-models.

parameters
Regular expression pattern that describes the parameters that should be returned. Meta-parameters (like lp__ or prior_) are filtered by default, so only parameters that typically appear in the summary() are returned. Use parameters to select specific parameters for the output.

Value

A data frame containing the lower and upper bounds of the SI.
Note that if the level of requested support is higher than observed in the data, the interval will be [NA,NA].

Details

This method is used to compute support intervals based on prior and posterior distributions. For the computation of support intervals, the model priors must be proper priors (at the very least they should be not flat, and it is preferable that they be informative - note that by default, brms::brm() uses flat priors for fixed-effects; see example below).

Setting the correct prior

It is important to provide the correct prior for meaningful results.

When posterior is a numerical vector, prior should also be a numerical vector.

When posterior is a data.frame, prior should also be a data.frame, with matching column order.

When posterior is a stanreg or brmsfit model:

prior can be set to NULL, in which case prior samples are drawn internally.

prior can also be a model equvilant to posterior but with samples from the priors only.

When posterior is an emmGrid object:

prior should be the stanreg or brmsfit model used to create the emmGrid objects.

prior can also be an emmGrid object equvilant to posterior but created with a model of priors samples only.

Choosing a value of BF

The choice of BF (the level of support) depends on what we want our interval to represent:

A BF = 1 contains values whose credibility is not decreased by observing the data.

A BF > 1 contains values who recived more impressive support from the data.

A BF < 1 contains values whose credibility has not been impressively decreased by observing the data. Testing against values outside this interval will produce a Bayes factor larger than 1/BF in support of the alternative. E.g., if an SI (BF = 1/3) excludes 0, the Bayes factor against the point-null will be larger than 3.

References

Wagenmakers, E., Gronau, Q. F., Dablander, F., & Etz, A. (2018, November 22). The Support Interval. doi: 10.31234/osf.io/zwnxb

Examples

library(bayestestR)

prior <- distribution_normal(1000, mean = 0, sd = 1)
posterior <- distribution_normal(1000, mean = .5, sd = .3)

si(posterior, prior)

#> # Support Interval
#> 
#> BF = 1 SI   
#> ------------
#> [0.04, 1.05]
#> 

if (FALSE) {
# rstanarm models
# ---------------
library(rstanarm)
contrasts(sleep$group) <- contr.bayes # see vingette
stan_model <- stan_lmer(extra ~ group + (1 | ID), data = sleep)
si(stan_model)
si(stan_model, BF = 3)

# emmGrid objects
# ---------------
library(emmeans)
group_diff <- pairs(emmeans(stan_model, ~group))
si(group_diff, prior = stan_model)

# brms models
# -----------
library(brms)
contrasts(sleep$group) <- contr.bayes # see vingette
my_custom_priors <-
  set_prior("student_t(3, 0, 1)", class = "b") +
  set_prior("student_t(3, 0, 1)", class = "sd", group = "ID")

brms_model <- brm(extra ~ group + (1 | ID),
  data = sleep,
  prior = my_custom_priors
)
si(brms_model)
}